### 2025-08-14 #8804 @editorialbot
[PRE REVIEW]: Open, Reproducible Calculation of Assembly Indices
**Submitting author:** @jdaymude (<a href="http://orcid.org/0000-0001-7294-5626">Joshua J. Daymude</a>)
**Repository:** https://github.com/DaymudeLab/assembly-theory
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.5.1
