### 2025-08-15 7bc6c10
Add python file for plotting search tree info
