+++ scripts/search_tree.py
@@ -1,0 +1,191 @@
+import cbor2
+import os
+import matplotlib.pyplot as plt
+import itertools
+
+class SearchNode:
+    def __init__(self, bound_list):
+        self.bounds = set()
+        self.children = []
+
+        for b in bound_list:
+            self.bounds.add(b)
+
+    # Given cbor tree file, return root of SearchNode tree
+    def from_cbor(tree_cbor):
+        parent = SearchNode(tree_cbor['bounds'])
+
+        for node in tree_cbor['children']:
+            temp = SearchNode.from_cbor(node)
+            parent.children.append(temp)
+
+        return parent
+    
+    # Given a list of bounds, tell the number of states searched
+    # with these bounds enabled
+    # Bounds can be Log, Int, VecSimple, VecSmallFrags
+    def num_searched(self, bounds):
+        if len(bounds & self.bounds) > 0:
+            return 1
+        else:
+            return 1 + sum([x.num_searched(bounds) for x in self.children])
+        
+    # Prove score given bounds and their timing information
+    # bounds_weights: [(bound, weight)]
+    def score(self, bounds_weights):
+        total_weight = 0
+        for b, w in bounds_weights:
+            total_weight += w
+            if b in self.bounds:
+                return total_weight
+        
+        return total_weight + sum([x.score(bounds_weights) for x in self.children])
+
+class Timer:
+    def __init__(self, timer, bound_name):
+        # timer: {metric -> [{time_unit -> time}, num]}
+        self.timer = timer
+        self.bound = bound_name
+
+    def from_cbor(name, timer_data):
+        name_map = {
+            'Log': 'log_timer',
+            'Int': 'int_timer',
+            'VecSimple': 'vec_simple_timer',
+            'VecSmallFrags': 'vec_small_frags_timer',
+            'Memoize': 'memoize_timer',
+        }
+
+        bound_name = name_map[name]
+        return Timer(timer_data[bound_name], name)
+
+    # Average time of the timer in nanos
+    def average(self):
+        total_time = 0
+        total_num = 0
+        for val in self.timer.values():
+            total_time += 1e9 * val[0]['secs'] + val[0]['nanos']
+            total_num += val[1]
+
+        return total_time / total_num
+
+def print_average_bound_times(dataset):
+    timer_strings = ['Log', 'Int', 'VecSimple', 'VecSmallFrags']
+
+    total_avg = {name: 0 for name in timer_strings}
+    num_mols = 0
+
+    for mol_name in os.listdir(dataset):
+        num_mols += 1
+
+        timer_file = open(dataset + mol_name + '/timer.cbor', 'rb')
+        timers = cbor2.load(timer_file)
+        timer_file.close()
+
+        for name in timer_strings:
+            timer = Timer.from_cbor(name, timers)
+            avg = timer.average()
+            total_avg[name] += avg
+
+    for key, val in total_avg.items():
+        print(key, ":", val / num_mols)
+
+def output_states_searched(dataset, outfile):
+    # list of mols to gather data for
+    mols = os.listdir(dataset)
+    # Total states searched for each mol
+    total_search = []
+    # Total pruned for each bound
+    # Each bound also includes all previous
+    ordered_bounds = {
+        'Int': [],
+        'VecSimlpe': [],
+        'VecSmallFrags': [],
+    }
+
+    for mol in mols:
+        print(mol)
+        f = open(dataset + mol + '/tree.cbor', 'rb')
+        data = cbor2.load(f)
+        f.close()
[…]
