### 2025-08-15 b8ac013
Add timing information to nodes of search tree
