### 2025-08-15 8ee945c
Python no longer uses intermediate data structure for scoring
