+++ scripts/search_tree.py
@@ -2,24 +2,41 @@ import cbor2
 import os
 import matplotlib.pyplot as plt
 import itertools
+from matplotlib_venn import venn3
 
 class SearchNode:
-    def __init__(self, bound_list):
+    def __init__(self, bound_list, dur_list):
         self.bounds = set()
+        self.times = list()
         self.children = []
 
         for b in bound_list:
             self.bounds.add(b)
 
+        for dur in dur_list:
+            nanos = dur['secs'] * 1e9 + dur['nanos']
+            self.times.append(nanos)
+
     # Given cbor tree file, return root of SearchNode tree
     def from_cbor(tree_cbor):
-        parent = SearchNode(tree_cbor['bounds'])
+        parent = SearchNode(tree_cbor['bounds'], tree_cbor['times'])
 
         for node in tree_cbor['children']:
             temp = SearchNode.from_cbor(node)
             parent.children.append(temp)
 
         return parent
+    
+    def bound_time(self, bound):
+        index = 0
+        match bound:
+            case 'Log': index = 0
+            case 'Int': index = 1
+            case 'VecSimple': index = 2
+            case 'VecSmallFrag': index = 3
+            case 'Memoize': index = 4
+        
+        return self.times[index]
     
     # Given a list of bounds, tell the number of states searched
     # with these bounds enabled
@@ -30,16 +47,15 @@ class SearchNode:
         else:
             return 1 + sum([x.num_searched(bounds) for x in self.children])
         
-    # Prove score given bounds and their timing information
-    # bounds_weights: [(bound, weight)]
-    def score(self, bounds_weights):
-        total_weight = 0
-        for b, w in bounds_weights:
-            total_weight += w
+    # Prove score given bounds
+    def score(self, bounds):
+        total_weight = 500
+        for b in bounds:
+            total_weight += self.bound_time(b)
             if b in self.bounds:
                 return total_weight
         
-        return total_weight + sum([x.score(bounds_weights) for x in self.children])
+        return total_weight + sum([x.score(bounds) for x in self.children])
 
 class Timer:
     def __init__(self, timer, bound_name):
@@ -99,7 +115,7 @@ def output_states_searched(dataset, outfile):
     # Each bound also includes all previous
     ordered_bounds = {
         'Int': [],
-        'VecSimlpe': [],
+        'VecSimple': [],
         'VecSmallFrags': [],
     }
 
@@ -161,31 +177,44 @@ def plot_states_searched(infile):
     plt.show()
 
 def score_mol(mol_file, bounds):
-    f = open(mol_file + '/timer.cbor', 'rb')
+    '''f = open(mol_file + '/timer.cbor', 'rb')
     timer_data = cbor2.load(f)
     f.close()
-    timers = [Timer.from_cbor(b, timer_data) for b in bounds]
+    timers = [Timer.from_cbor(b, timer_data) for b in bounds]'''
 
     f = open(mol_file + '/tree.cbor', 'rb')
     tree_data = cbor2.load(f)
     f.close()
-    root = SearchNode.from_cbor(tree_data)
 
-    bounds_weights = [(x.bound, x.average()) for x in timers]
     scores = []
-    for n in range(1, len(bounds_weights) + 1):
-        for perm in itertools.permutations(bounds_weights, n):
-            bounds = [x[0] for x in perm]
-            score = root.score(perm)
-            scores.append((bounds, score))
+    for n in range(1, len(bounds) + 1):
+        for perm in itertools.permutations(bounds, n):
+            score = score_cbor(tree_data, perm)
+            scores.append((perm, score))
     
-    scores.sort(key=lambda x: -x[1])
-    for s in scores:
[…]
