+++ scripts/search_tree.py
@@ -3,6 +3,9 @@ import os
 import matplotlib.pyplot as plt
 import itertools
 from matplotlib_venn import venn3
+from matplotlib_venn.layout.venn3 import DefaultLayoutAlgorithm
+import numpy as np
+import gc
 
 class SearchNode:
     def __init__(self, bound_list, dur_list):
@@ -112,12 +115,9 @@ def output_states_searched(dataset, outfile):
     # Total states searched for each mol
     total_search = []
     # Total pruned for each bound
-    # Each bound also includes all previous
-    ordered_bounds = {
-        'Int': [],
-        'VecSimple': [],
-        'VecSmallFrags': [],
-    }
+    pruned = dict()
+
+    bounds = ['Int', 'VecSimple', 'VecSmallFrags']
 
     for mol in mols:
         print(mol)
@@ -126,16 +126,25 @@ def output_states_searched(dataset, outfile):
         f.close()
         root = SearchNode.from_cbor(data)
 
-        bounds_set = set()
-        states_total = root.num_searched(bounds_set)
+        states_total = root.num_searched(set())
         total_search.append(states_total)
-        for bound in ordered_bounds:
-            bounds_set.add(bound)
-            ordered_bounds[bound].append(states_total - root.num_searched(bounds_set))
+
+        for n in range(1, len(bounds) + 1):
+            for comb in itertools.combinations(bounds, n):
+                num_pruned = states_total - root.num_searched(set(comb))
+                if comb in pruned:
+                    pruned[comb].append(num_pruned)
+                else:
+                    pruned[comb] = [num_pruned]
+
+        del data
+        gc.collect()
 
     with open(outfile, 'wb') as f:
-        cbor2.dump({'mols': mols, 'totals': total_search, 'pruned': ordered_bounds}, f)
+        cbor2.dump({'mols': mols, 'totals': total_search, 'pruned': pruned}, f)
 
+'''
+Broken for now
 def plot_states_searched(infile):
     f = open(infile, 'rb')
     data = cbor2.load(f)
@@ -173,8 +182,8 @@ def plot_states_searched(infile):
     plt.xticks([])
     plt.legend(loc="upper left")
     plt.xlabel('Molecule')
-    plt.ylabel('% States Pruned')
-    plt.show()
+    plt.ylabel('# states pruned / # states searched (log-only)')
+    plt.show()'''
 
 def score_mol(mol_file, bounds):
     '''f = open(mol_file + '/timer.cbor', 'rb')
@@ -218,3 +227,53 @@ def score_cbor(node, bounds):
         
     return total_weight + 500 + sum([score_cbor(x, bounds) for x in node['children']])
 
+def venn_diagram(inputfile):
+    # input file is generated by output_states_searched
+    f = open(inputfile, 'rb')
+    data = cbor2.load(f)
+    f.close()
+
+    mols = data['mols']
+    total_search = data['totals']
+    pruned = data['pruned']
+
+    num_mols = len(mols)
+    # Create matrix to go from states_searched output to 
+    # venn diagram values
+    include_mat = np.array([
+        [1, 0, 1, 0, 1, 0, 1],
+        [0, 1, 1, 0, 0, 1, 1],
+        [1, 1, 1, 0, 1, 1, 1],
+        [0, 0, 0, 1, 1, 1, 1],
+        [1, 0, 1, 1, 1, 1, 1],
+        [0, 1, 1, 1, 1, 1, 1],
+        [1, 1, 1, 1, 1, 1, 1],
+    ])
+    include_mat = np.linalg.inv(include_mat)
+
[…]
