### 2025-08-20 a41e325
Add optimized matches generation
