### 2025-08-20 77f2166
Add non-bound time to search nodes
