+++ scripts/search_tree.py
@@ -185,13 +185,13 @@ def plot_states_searched(infile):
     plt.ylabel('# states pruned / # states searched (log-only)')
     plt.show()'''
 
-def score_mol(mol_file, bounds):
+def score_mol(mol_dir, bounds):
     '''f = open(mol_file + '/timer.cbor', 'rb')
     timer_data = cbor2.load(f)
     f.close()
     timers = [Timer.from_cbor(b, timer_data) for b in bounds]'''
 
-    f = open(mol_file + '/tree.cbor', 'rb')
+    f = open(mol_dir + 'tree.cbor', 'rb')
     tree_data = cbor2.load(f)
     f.close()
 
@@ -208,6 +208,7 @@ def score_cbor(node, bounds):
     total_weight = 0
     node_bounds = node['bounds']
     node_times = node['times']
+    default_weight = node_times[5]['secs'] * 1e9 + node_times[5]['nanos']
 
     for b in bounds:
         index = 0
@@ -224,8 +225,8 @@ def score_cbor(node, bounds):
         total_weight += bound_time
         if b in node_bounds:
             return total_weight
-        
-    return total_weight + 500 + sum([score_cbor(x, bounds) for x in node['children']])
+
+    return total_weight + default_weight + sum([score_cbor(x, bounds) for x in node['children']])
 
 def venn_diagram(inputfile):
     # input file is generated by output_states_searched
@@ -275,5 +276,6 @@ def venn_diagram(inputfile):
     plt.show()
 
 
-#output_states_searched('eval_out/coconut_55/', 'eval_out/coconut_55_venndata')
-venn_diagram('eval_out/coconut_55_venndata')
+scores = score_mol('eval_out/gdb17_200/002.mol/', ['Int', 'VecSimple', 'VecSmallFrags'])
+for s in scores:
+    print(s)
