### 2025-08-21 784b2fc
Add bound order heatmap plot
