### 2025-08-21 fe50c12
Add bound order plotting for datasets
