### 2025-08-21 26cea2f
Add new int bound
