+++ src/assembly.rs
@@ -193,6 +193,19 @@ fn fragments(mol: &Molecule, state: &[BitSet], h1: &BitSet, h2: &BitSet) -> Opti
     Some(fragments)
 }
 
+fn is_usable(state: &[BitSet], h1: &BitSet, h2: &BitSet, masks: &mut Vec<Vec<BitSet>>) -> bool {
+    let f1 = state.iter().enumerate().find(|(_, c)| h1.is_subset(c));
+    let f2 = state.iter().enumerate().find(|(_, c)| h2.is_subset(c));
+    
+    if let (Some((i1, _)), Some((i2, _))) = (f1, f2) {
+        masks[h1.len() - 2][i1].union_with(h1);
+        masks[h1.len() - 2][i2].union_with(h2);
+        true
+    } else {
+        false
+    }
+}
+
 /// Recursive helper for [`index_search`], only public for benchmarking.
 ///
 /// Inputs:
@@ -230,6 +243,9 @@ pub fn recurse_index_search(
     parallel_mode: ParallelMode,
     kernel_mode: KernelMode,
 ) -> (usize, usize) {
+    //println!("{:?}", removal_order);
+    //println!("{:?}", state);
+
     // An upper bound on the size of fragments that can be
     // removed from this or any descendant assembly state.
     let largest_remove = {
@@ -241,19 +257,40 @@ pub fn recurse_index_search(
         }
     };
 
+    let mut masks: Vec<Vec<BitSet>> = vec![vec![BitSet::with_capacity(mol.graph().edge_count()); state.len()]; largest_remove - 1];
+    for v in subgraph.iter() {
+        let m = &matches[v];
+        is_usable(state, &m.0, &m.1, &mut masks);
+    }
+    /*while masks.iter().last().unwrap().len() == 0 {
+        masks.pop();
+    }*/
+    /*for (i, m) in masks[0].iter().enumerate() {
+        state[i] = m.clone();
+    }*/
+    /*let mut state = Vec::new();
+    for m in masks[0].iter() {
+        if m.len() != 0 {
+            state.extend(connected_components_under_edges(mol.graph(), m));
+        }
+    }
+    state.retain(|i| i.len() > 1);*/
+
+
     // If any bounds would prune this assembly state or if memoization is
     // enabled and this assembly state is preempted by the cached state, halt.
     if bound_exceeded(
         mol,
         matches,
         graph,
-        state,
+        &state,
         &subgraph,
         state_index,
         best_index.load(Relaxed),
         largest_remove,
         bounds,
-    ) || cache.memoize_state(mol, state, state_index, &removal_order)
+        &masks,
+    ) || cache.memoize_state(mol, &state, state_index, &removal_order)
     {
         return (state_index, 1);
     }
@@ -281,7 +318,7 @@ pub fn recurse_index_search(
     // the given (enumerated) pair of non-overlapping isomorphic subgraphs.
     let recurse_on_match = |v: usize| {
         let (h1, h2) = &matches[v];
-        if let Some(fragments) = fragments(mol, state, h1, h2) {
+        if let Some(fragments) = fragments(mol, &state, h1, h2) {
             // If using depth-one parallelism, all descendant states should be
             // computed serially.
             let new_parallel = if parallel_mode == ParallelMode::DepthOne {
@@ -482,6 +519,8 @@ pub fn index_search(
         parallel_mode,
         kernel_mode,
     );
+
+    println!("{}", cache.count());
 
     (index as u32, matches.len() as u32, states_searched)
 }
