### 2025-08-21 f0c38bc
Add cache counter
