+++ src/assembly.rs
@@ -19,7 +19,7 @@
 //! ```
 
 use std::{
-    collections::HashMap, sync::{
+    collections::{HashMap, HashSet}, sync::{
         atomic::{AtomicUsize, Ordering::Relaxed},
         Arc,
     }
@@ -28,15 +28,16 @@ use std::{
 use bit_set::BitSet;
 use clap::ValueEnum;
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
+use petgraph::graph::EdgeIndex;
 
 use crate::{
     bounds::{bound_exceeded, Bound},
     canonize::{canonize, CanonizeMode, Labeling},
-    enumerate::{enumerate_subgraphs, EnumerateMode},
+    enumerate::{EnumerateMode},
     kernels::{KernelMode, deletion_kernel, inclusion_kernel},
     memoize::{Cache, MemoizeMode},
     molecule::Molecule,
-    utils::connected_components_under_edges,
+    utils::{connected_components_under_edges, edge_neighbors},
     reductions::CompatGraph,
 };
 
@@ -98,38 +99,82 @@ pub fn depth(mol: &Molecule) -> u32 {
 ///
 /// Return all pairs of non-overlapping, isomorphic subgraphs in the molecule,
 /// sorted to guarantee deterministic iteration.
-pub fn matches(
-    mol: &Molecule,
-    enumerate_mode: EnumerateMode,
-    canonize_mode: CanonizeMode,
-) -> Vec<(BitSet, BitSet)> {
-    // Enumerate all connected, non-induced subgraphs with at most |E|/2 edges
-    // and bin them into isomorphism classes using canonization.
-    let mut isomorphism_classes = HashMap::<Labeling, Vec<BitSet>>::new();
-    for subgraph in enumerate_subgraphs(mol, enumerate_mode) {
-        isomorphism_classes
-            .entry(canonize(mol, &subgraph, canonize_mode))
-            .and_modify(|bucket| bucket.push(subgraph.clone()))
-            .or_insert(vec![subgraph.clone()]);
+pub fn matches(mol: &Molecule) -> Vec<(BitSet, BitSet)> {
+    let num_edges = mol.graph().edge_count();
+    let mut subgraphs: HashSet<BitSet> = HashSet::new();
+    let mut size_one_subgraphs: HashSet<BitSet> = HashSet::new();
+    let mut matches: Vec<(BitSet, BitSet)> = Vec::new();
+    let mut buckets: HashMap<Labeling, Vec<BitSet>> = HashMap::new();
+
+    // Generate subgraphs with one edge
+    for i in 0..num_edges {
+        let mut new_bitset = BitSet::with_capacity(num_edges);
+        new_bitset.insert(i);
+        size_one_subgraphs.insert(new_bitset);
     }
 
-    // In each isomorphism class, collect non-overlapping pairs of subgraphs.
-    let mut matches = Vec::new();
-    for bucket in isomorphism_classes.values() {
-        for (i, first) in bucket.iter().enumerate() {
-            for second in &bucket[i..] {
-                if first.is_disjoint(second) {
-                    if first > second {
+    // Generate subgraphs with two edges
+    for subgraph in size_one_subgraphs.iter() {
+        let mut neighborhood = BitSet::with_capacity(num_edges);
+        for e in subgraph {
+            neighborhood.extend(edge_neighbors(mol.graph(), EdgeIndex::new(e)).map(|x| x.index()));
+        }
+        neighborhood.difference_with(subgraph);
+
+        for e in neighborhood.iter() {
+            let mut new_subgraph = subgraph.clone();
+            new_subgraph.insert(e);
+            subgraphs.insert(new_subgraph);
+        }
+    }
+
[…]
