### 2025-08-21 fef2d13
Remove cache counter
