### 2025-08-26 ccd5a79
new memoize strategy
