+++ src/main.rs
@@ -122,6 +122,47 @@ fn main() -> Result<()> {
         }
     }
 
+
+    /*use std::ffi::{OsStr, OsString};
+    use std::path::Path;
+    use assembly_theory::molecule::Molecule;
+    use rayon::iter::{IntoParallelIterator, ParallelIterator, IntoParallelRefIterator, IndexedParallelIterator};
+
+    let paths = fs::read_dir(Path::new("data").join("plycyclic_hydrocarbons")).unwrap();
+    let mut mol_list: Vec<Molecule> = Vec::new();
+    let mut names: Vec<OsString> = Vec::new();
+
+    for path in paths {
+        let name = path.unwrap().path();
+        if name.extension().and_then(OsStr::to_str) != Some("mol") {
+            continue;
+        }
+        names.push(name.file_name().unwrap().to_os_string());
+        mol_list.push(
+            parse_molfile_str(
+                &fs::read_to_string(name.clone())
+                    .expect(&format!("Could not read file {name:?}")),
+            )
+            .expect(&format!("Failed to parse {name:?}")),
+        );
+    }
+
+    names.par_iter().zip(mol_list.par_iter()).for_each(|(n, mol)| {
+        let index = index_search(
+            &mol,
+            cli.enumerate,
+            cli.canonize,
+            cli.parallel,
+            cli.memoize,
+            cli.kernel,            
+            boundlist,
+            cli.clique,
+        );
+        println!("{:?}: MA: {} Space: {}, Searched: {}", n, index.0, index.1, index.2);
+    });
+    std::process::exit(1);*/
+
+
     // Call index calculation with all the various options.
     let (index, num_matches, states_searched) = index_search(
         &mol,
