### 2025-08-26 4f0f7c0
Modified int bound
