### 2025-08-27 7bd9fdd
New (unoptimized) method of calculating masks
