+++ src/assembly.rs
@@ -128,6 +128,7 @@ pub fn matches(mol: &Molecule) -> Vec<(BitSet, BitSet)> {
         }
     }
 
+    // Generate subgraphs with more than two edges
     while subgraphs.len() > 0 {
         buckets.clear();
         // Generate buckets
@@ -175,6 +176,7 @@ pub fn matches(mol: &Molecule) -> Vec<(BitSet, BitSet)> {
         }
     }
 
+    // Sort matches
     matches.sort_by(|e1, e2| {
         let ord = [
             e2.0.len().cmp(&e1.0.len()), // Decreasing subgraph size.
@@ -291,8 +293,6 @@ pub fn recurse_index_search(
     //println!("{:?}", removal_order);
     //println!("{:?}", state);
 
-    // An upper bound on the size of fragments that can be
-    // removed from this or any descendant assembly state.
     let largest_remove = {
         if let Some(v) = subgraph.iter().next() {
             matches[v].0.len()
@@ -302,47 +302,63 @@ pub fn recurse_index_search(
         }
     };
 
+    // Prune edges of state
+    let mut subgraphs: HashSet<BitSet> = HashSet::new();
     let mut masks: Vec<Vec<BitSet>> = vec![vec![BitSet::with_capacity(mol.graph().edge_count()); state.len()]; largest_remove - 1];
-    for v in subgraph.iter() {
-        let m = &matches[v];
-        is_usable(state, &m.0, &m.1, &mut masks);
-    }
+    subgraph = subgraph.iter().filter(|v| {
+        let h1 = &matches[*v].0;
+        let h2 = &matches[*v].1;
+
+        let pre_have_h1 = subgraphs.contains(h1);
+        let pre_have_h2 = subgraphs.contains(h2);
+
+        let mut have_h1 = pre_have_h1;
+        let mut have_h2 = pre_have_h2;
+
+        let mut idx1 = 0;
+        let mut idx2 = 0;
+
+        if !pre_have_h1 {
+            if let Some((idx, _)) = state.iter().enumerate().find(|(_, frag)| h1.is_subset(frag)) {
+                idx1 = idx;
+                have_h1 = true;
+            }
+        }
+        if !pre_have_h2 {
+            if let Some((idx, _)) = state.iter().enumerate().find(|(_, frag)| h2.is_subset(frag)) {
+                idx2 = idx;
+                have_h2 = true;
+            }
+        }
+
+        if have_h1 && have_h2 {
+            if !pre_have_h1 {
+                subgraphs.insert(h1.clone());
+                masks[h1.len() - 2][idx1].union_with(h1);
+            }
+            if !pre_have_h2 {
+                subgraphs.insert(h2.clone());
+                masks[h1.len() - 2][idx2].union_with(h2);
+            }
+        }
+
+        have_h1 && have_h2
+    }).collect();
+
+    /*let mut masks: Vec<Vec<BitSet>> = vec![vec![BitSet::with_capacity(mol.graph().edge_count()); state.len()]; largest_remove - 1];
+    let subgraph1 = subgraph.iter().filter(|v| {
+        let m = &matches[*v];
+        is_usable(state, &m.0, &m.1, &mut masks)
+    }).collect::<BitSet>();*/
 
     let mut state = Vec::new();
-    for (i, m) in masks[0].iter().enumerate() {
+    for m in masks[0].iter() {
         if m.len() != 0 {
             state.extend(connected_components_under_edges(mol.graph(), m));
         }
     }
-    /*while masks.iter().last().unwrap().len() == 0 {
-        masks.pop();
-    }*/
-    /*let mut state = Vec::new();
-    for (i, m) in masks[0].iter().enumerate() {
[…]
