### 2025-08-28 66256cd
Optimize masks computation
