+++ src/assembly.rs
@@ -29,14 +29,14 @@ use bit_set::BitSet;
 use clap::ValueEnum;
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 use petgraph::graph::EdgeIndex;
-use ahash::{AHasher, RandomState};
+use ahash::RandomState;
 
 use crate::{
     bounds::{bound_exceeded, Bound},
     canonize::{canonize, CanonizeMode, Labeling},
     enumerate::{EnumerateMode},
     kernels::{KernelMode, deletion_kernel, inclusion_kernel},
-    memoize::{Cache, MemoizeMode, NewCache},
+    memoize::{MemoizeMode, NewCache},
     molecule::Molecule,
     utils::{connected_components_under_edges, edge_neighbors},
     reductions::CompatGraph,
@@ -100,22 +100,28 @@ pub fn depth(mol: &Molecule) -> u32 {
 ///
 /// Return all pairs of non-overlapping, isomorphic subgraphs in the molecule,
 /// sorted to guarantee deterministic iteration.
-pub fn matches(mol: &Molecule) -> Vec<(BitSet, BitSet)> {
+pub fn matches(mol: &Molecule) -> (Vec<(BitSet, BitSet)>, HashMap<usize, (BitSet, Vec<usize>)>) {
     let num_edges = mol.graph().edge_count();
-    let mut subgraphs: HashSet<BitSet> = HashSet::new();
-    let mut size_one_subgraphs: HashSet<BitSet> = HashSet::new();
+    let mut subgraphs: HashMap<BitSet, usize> = HashMap::new();
+    let mut size_one_subgraphs: HashMap<BitSet, usize> = HashMap::new();
     let mut matches: Vec<(BitSet, BitSet)> = Vec::new();
-    let mut buckets: HashMap<Labeling, Vec<BitSet>> = HashMap::new();
+    let mut buckets: HashMap<Labeling, Vec<(BitSet, usize)>> = HashMap::new();
+
+    let mut dag: HashMap<usize, (BitSet, Vec<usize>)> = HashMap::new();
+    let mut next_id = 0;
 
     // Generate subgraphs with one edge
     for i in 0..num_edges {
         let mut new_bitset = BitSet::with_capacity(num_edges);
         new_bitset.insert(i);
-        size_one_subgraphs.insert(new_bitset);
+        size_one_subgraphs.insert(new_bitset, i);
+
+        dag.insert(next_id, (BitSet::new(), Vec::new()));
+        next_id += 1;
     }
 
     // Generate subgraphs with two edges
-    for subgraph in size_one_subgraphs.iter() {
+    for (subgraph, idx) in size_one_subgraphs.iter() {
         let mut neighborhood = BitSet::with_capacity(num_edges);
         for e in subgraph {
             neighborhood.extend(edge_neighbors(mol.graph(), EdgeIndex::new(e)).map(|x| x.index()));
@@ -125,7 +131,8 @@ pub fn matches(mol: &Molecule) -> Vec<(BitSet, BitSet)> {
         for e in neighborhood.iter() {
             let mut new_subgraph = subgraph.clone();
             new_subgraph.insert(e);
-            subgraphs.insert(new_subgraph);
+            subgraphs.entry(new_subgraph)
+                .or_insert(*idx);
         }
     }
 
@@ -133,45 +140,68 @@ pub fn matches(mol: &Molecule) -> Vec<(BitSet, BitSet)> {
     while subgraphs.len() > 0 {
         buckets.clear();
         // Generate buckets
-        for g in subgraphs.iter() {
-            let canon = canonize(mol, g, CanonizeMode::TreeNauty);
+        for (g, parent) in subgraphs.iter() {
[…]
