+++ src/assembly.rs
@@ -29,7 +29,7 @@ use bit_set::BitSet;
 use clap::ValueEnum;
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 use petgraph::graph::EdgeIndex;
-use ahash::RandomState;
+use itertools::Itertools;
 
 use crate::{
     bounds::{bound_exceeded, Bound},
@@ -96,19 +96,212 @@ pub fn depth(mol: &Molecule) -> u32 {
     ix
 }
 
+#[derive(Debug)]
+pub struct DagNode {
+    fragment: BitSet,
+    canon_id: usize,
+    // added_edge: usize,
+    children: Vec<usize>,
+}
+
+impl DagNode {
+    pub fn new(fragment: BitSet, canon_id: usize) -> Self {
+        Self {
+            fragment,
+            canon_id,
+            children: Vec::new(),
+        }
+    }
+}
+
+
+pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (HashMap<(usize, usize), usize>, Vec<DagNode>) {
+    let num_edges = mol.graph().edge_count();
+
+    let mut subgraphs: Vec<usize> = Vec::new();
+    let mut constructed_frags: HashSet<BitSet> = HashSet::new();
+    let mut frag_to_id: HashMap<BitSet, usize> = HashMap::new();
+
+    let mut dag: Vec<DagNode> = Vec::with_capacity(num_edges);
+    let mut matches: Vec<(usize, usize)> = Vec::new();
+
+    let mut next_canon_id = 0;
+
+    // Generate subgraphs with one edge
+    for i in 0..num_edges {
+        let mut new_bitset = BitSet::with_capacity(num_edges);
+        new_bitset.insert(i);
+        
+        let node = DagNode::new(new_bitset, 0);
+
+        dag.push(node);
+        subgraphs.push(i);
+    }
+
+    next_canon_id += 1;
+
+    while subgraphs.len() > 0 {
+        // Initialize new buckets and subgraph list
+        let mut buckets: HashMap<Labeling, Vec<(BitSet, usize)>> = HashMap::new();
+        let mut child_subgraphs: Vec<usize> = Vec::new();
+
+        // Extend and bucket new subgraphs
+        for parent_idx in subgraphs.iter() {
+            // Get fragment to be extended
+            let fragment = &dag[*parent_idx].fragment;
+
+            // Construct edge neighborhood of fragment
+            let mut neighborhood = BitSet::with_capacity(num_edges);
+            for e in fragment {
+                neighborhood.extend(edge_neighbors(mol.graph(), EdgeIndex::new(e)).map(|x| x.index()));
+            }
+            neighborhood.difference_with(fragment);
+
+            // Extend fragment with edges from neighborhood
+            for e in neighborhood.iter() {
+                // Create extended fragment
+                let mut new_fragment = fragment.clone();
+                new_fragment.insert(e);
+
+                // Check if fragment has already been created
+                // If yes, continue
+                if constructed_frags.contains(&new_fragment) {
+                    continue;
+                }
+
+                // Bucket subgraph
+                // Gives ownership of new_fragment to buckets
+                let canon = canonize(mol, &new_fragment, canonize_mode);
+                buckets.entry(canon)
+                    .and_modify(|bucket| bucket.push((new_fragment.clone(), *parent_idx)))
+                    .or_insert(vec![(new_fragment.clone(), *parent_idx)]);
+            
+                // Add to constructed_frags
+                constructed_frags.insert(new_fragment);
+            }
+        }
+
+        // Go through buckets to create matches
+        for fragments in buckets.values() {
+            // Variables to track if a match has been found
+            let mut frag_has_match = BitSet::with_capacity(fragments.len());
+
[…]
