+++ Cargo.toml
@@ -20,6 +20,7 @@ rayon = "1.10.0"
 dashmap = "6.1.0"
 ahash = "0.8.12"
 fxhash = "0.2.1"
+itertools = "0.14.0"
 
 [dev-dependencies]
 criterion = "0.3"
