### 2025-09-03 e700ed6
Enumerate matches using dag
