+++ src/assembly.rs
@@ -19,7 +19,7 @@
 //! ```
 
 use std::{
-    collections::{HashMap, HashSet}, sync::{
+    collections::{BTreeMap, HashMap, HashSet}, sync::{
         atomic::{AtomicUsize, Ordering::Relaxed},
         Arc,
     }
@@ -27,9 +27,11 @@ use std::{
 
 use bit_set::BitSet;
 use clap::ValueEnum;
+use graph_canon::canon;
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 use petgraph::graph::EdgeIndex;
 use itertools::Itertools;
+use std::cmp::{max, min};
 
 use crate::{
     bounds::{bound_exceeded, Bound},
@@ -142,7 +144,7 @@ pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (HashMap<(usize, 
 
     while subgraphs.len() > 0 {
         // Initialize new buckets and subgraph list
-        let mut buckets: HashMap<Labeling, Vec<(BitSet, usize)>> = HashMap::new();
+        let mut buckets: BTreeMap<Labeling, Vec<(BitSet, usize)>> = BTreeMap::new();
         let mut child_subgraphs: Vec<usize> = Vec::new();
 
         // Extend and bucket new subgraphs
@@ -506,11 +508,10 @@ fn is_usable(state: &[BitSet], h1: &BitSet, h2: &BitSet, masks: &mut Vec<Vec<Bit
 ///   upper bound returned is unchanged.)
 /// - `usize`: The number of assembly states searched.
 #[allow(clippy::too_many_arguments)]
-/*pub fn recurse_index_search(
+pub fn recurse_index_search(
     mol: &Molecule,
-    matches: &Vec<(usize, usize)>,
-    dag: &HashMap<usize, (BitSet, Vec<usize>)>,
-    graph: &Option<CompatGraph>,
+    matches: &HashMap<(usize, usize), usize>,
+    dag: &Vec<DagNode>,
     mut subgraph: BitSet,
     removal_order: Vec<usize>,
     state: &[BitSet],
@@ -523,85 +524,144 @@ fn is_usable(state: &[BitSet], h1: &BitSet, h2: &BitSet, masks: &mut Vec<Vec<Bit
 ) -> (usize, usize) {
     //println!("{:?}", removal_order);
     //println!("{:?}", state);
+    
+    let num_edges = mol.graph().edge_count();
+    let mut enum_matches: Vec<(usize, usize)> = Vec::new();
+    
+    // MATCHES METHOD 1
+    // EXTEND ALWAYS AND DO MATCHING AT END
+    let mut subgraphs: Vec<usize> = Vec::new();
+    let mut buckets: HashMap<usize, Vec<usize>> = HashMap::new();
 
-    let largest_remove = {
-        if let Some(v) = subgraph.iter().next() {
-            dag.get(matches[v].0).unwrap().0.len();
+    // create subgraphs of size 1
+    for fragment in state.iter() {
+        for edge_idx in fragment.iter() {
+            subgraphs.push(edge_idx);
         }
-        else {
-            return (state_index, 1);
-        }
-    };
+    }
 
-    let mut masks: Vec<Vec<BitSet>> = vec![
-        vec![BitSet::with_capacity(mol.graph().edge_count()); state.len()]; 
+    while subgraphs.len() > 0 {
+        // Subgraphs to extend in next loop
+        let mut new_subgraphs = Vec::new();
+
+        // Extend each subgraph and bucket
+        for frag_id in subgraphs {
+            // Get ids of extensions of this subgraph
+            let children_ids = &dag[frag_id].children;
+
+            for child_id in children_ids {
+                // Check if this extension is valid
+                // TODO: save fragment id
+                // TODO: save edge id in DAG
[…]
