### 2025-09-03 0997b00
Create masks with frag_id to state_id map
