+++ src/assembly.rs
@@ -28,10 +28,11 @@ use std::{
 use bit_set::BitSet;
 use clap::ValueEnum;
 use graph_canon::canon;
-use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
+use rayon::iter::{IntoParallelIterator, IntoParallelRefIterator, ParallelIterator, IndexedParallelIterator};
 use petgraph::graph::EdgeIndex;
 use itertools::Itertools;
 use std::cmp::{max, min};
+use ahash::RandomState;
 
 use crate::{
     bounds::{bound_exceeded, Bound},
@@ -102,7 +103,6 @@ pub fn depth(mol: &Molecule) -> u32 {
 pub struct DagNode {
     fragment: BitSet,
     canon_id: usize,
-    // added_edge: usize,
     children: Vec<usize>,
 }
 
@@ -512,6 +512,7 @@ pub fn recurse_index_search(
     mol: &Molecule,
     matches: &HashMap<(usize, usize), usize>,
     dag: &Vec<DagNode>,
+    last_removed: usize,
     mut subgraph: BitSet,
     removal_order: Vec<usize>,
     state: &[BitSet],
@@ -525,18 +526,18 @@ pub fn recurse_index_search(
     //println!("{:?}", removal_order);
     //println!("{:?}", state);
     
-    let num_edges = mol.graph().edge_count();
-    let mut enum_matches: Vec<(usize, usize)> = Vec::new();
-    
     // MATCHES METHOD 1
     // EXTEND ALWAYS AND DO MATCHING AT END
-    let mut subgraphs: Vec<usize> = Vec::new();
-    let mut buckets: HashMap<usize, Vec<usize>> = HashMap::new();
+    let num_edges = mol.graph().edge_count();
+
+    let mut enum_matches: Vec<(usize, usize)> = Vec::new();
+    let mut subgraphs: Vec<(usize, usize)> = Vec::new();
+    let mut buckets: HashMap<usize, Vec<(usize, usize)>> = HashMap::new();
 
     // create subgraphs of size 1
-    for fragment in state.iter() {
+    for (state_id, fragment) in state.iter().enumerate() {
         for edge_idx in fragment.iter() {
-            subgraphs.push(edge_idx);
+            subgraphs.push((edge_idx, state_id));
         }
     }
 
@@ -545,27 +546,29 @@ pub fn recurse_index_search(
         let mut new_subgraphs = Vec::new();
 
         // Extend each subgraph and bucket
-        for frag_id in subgraphs {
+        for (frag_id, state_id) in subgraphs {
+            // Get state frament that this subgraph is contained in
+            let state_frag = &state[state_id];
+
             // Get ids of extensions of this subgraph
             let children_ids = &dag[frag_id].children;
 
             for child_id in children_ids {
                 // Check if this extension is valid
-                // TODO: save fragment id
-                // TODO: save edge id in DAG
+                // i.e. the extended subgraph is contained in a fragment of state
                 let child_frag = &dag[*child_id].fragment;
-                let valid = state.iter().any(|frag| child_frag.is_subset(frag));
+                let valid = child_frag.is_subset(state_frag);
 
                 if valid {
                     let canon_id = &dag[*child_id].canon_id;
                     
                     // Add fragment to bucket
                     buckets.entry(*canon_id)
-                        .and_modify(|bucket| bucket.push(*child_id))
-                        .or_insert(vec![*child_id]);
+                        .and_modify(|bucket| bucket.push((*child_id, state_id)))
+                        .or_insert(vec![(*child_id, state_id)]);
                     
                     // Add fragment to new subgraphs
-                    new_subgraphs.push(*child_id);
[…]
