### 2025-09-03 87ce547
Create masks by searching through state
