### 2025-09-03 e66d911
Add int bound
