### 2025-09-03 a14e9ed
Cleanup some code
