+++ src/assembly.rs
@@ -116,7 +116,10 @@ impl DagNode {
     }
 }
 
-
+/// Helper function for [`index_search`]; only public for benchmarking.
+///
+/// Return all pairs of non-overlapping, isomorphic subgraphs in the molecule,
+/// sorted to guarantee deterministic iteration.
 pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (HashMap<(usize, usize), usize>, Vec<DagNode>) {
     let num_edges = mol.graph().edge_count();
 
@@ -285,146 +288,6 @@ pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (HashMap<(usize, 
     (match_to_id, dag)
 }
 
-
-/// Helper function for [`index_search`]; only public for benchmarking.
-///
-/// Return all pairs of non-overlapping, isomorphic subgraphs in the molecule,
-/// sorted to guarantee deterministic iteration.
-/*pub fn _matches(mol: &Molecule) -> (Vec<(usize, usize)>, DagNode) {
-    let num_edges = mol.graph().edge_count();
-
-    let mut subgraphs: HashMap<BitSet, usize> = HashMap::new();
-    let mut size_one_subgraphs: HashMap<BitSet, usize> = HashMap::new();
-    let mut buckets: HashMap<Labeling, Vec<DagNode>> = HashMap::new();
-
-    let mut dag: Vec<DagNode> = Vec::with_capacity(num_edges);
-    let mut canon_ids: HashMap<Labeling, usize> = HashMap::new();
-    let mut matches: HashMap<(usize, usize), usize> = Hashmap::new();
-
-    let mut next_frag_id = 1;
-    let mut next_match_id = 0;
-    let mut next_canon_id = 1;
-
-    // Generate subgraphs with one edge
-    for i in 0..num_edges {
-        let mut new_bitset = BitSet::with_capacity(num_edges);
-        new_bitset.insert(i);
-        size_one_subgraphs.insert(new_bitset, i);
-
-        dag.push(DagNode::new(new_bitset, 0, 0));
-    }
-
-    // Generate subgraphs with two edges
-    for (subgraph, idx) in size_one_subgraphs.iter() {
-        let mut neighborhood = BitSet::with_capacity(num_edges);
-        for e in subgraph {
-            neighborhood.extend(edge_neighbors(mol.graph(), EdgeIndex::new(e)).map(|x| x.index()));
-        }
-        neighborhood.difference_with(subgraph);
-
-        for e in neighborhood.iter() {
-            let mut new_subgraph = subgraph.clone();
-            new_subgraph.insert(e);
-            subgraphs.entry(new_subgraph)
-                .or_insert(*idx);
-        }
-    }
-
-    // Generate subgraphs with more than two edges
-    while subgraphs.len() > 0 {
-        buckets.clear();
-        // Generate buckets
-        for (g, parent) in subgraphs.iter() {
-            let canon = canonize(mol, g, CanonizeMode::TreeNauty);            
-            buckets.entry(canon)
-                .and_modify(|v| v.push((g.clone(), *parent)))
-                .or_insert(vec![(g.clone(), *parent)]);
-        }
-
-        subgraphs.clear();
-        let mut frag_to_id: HashMap<BitSet, usize> = HashMap::new();
-        for (_, bucket) in buckets.iter() {
-            // Generate matches
-            let mut has_match = BitSet::with_capacity(bucket.len());
-            for (i, (first, first_parent)) in bucket.iter().enumerate() {
-                for (j, (second, second_parent)) in bucket[i+1..].iter().enumerate() {
-                    if first.intersection(second).count() == 0 {
-                        if !has_match.contains(i) {
-                            frag_to_id.insert(first.clone(), next_id);
-
[…]
