### 2025-09-03 b2689d8
Don't extend subgraphs without a match
