+++ src/assembly.rs
@@ -22,7 +22,7 @@ use std::{
     collections::{BTreeMap, HashMap, HashSet}, sync::{
         atomic::{AtomicUsize, Ordering::Relaxed},
         Arc,
-    }
+    }, time::{Duration, Instant}
 };
 
 use bit_set::BitSet;
@@ -33,6 +33,7 @@ use petgraph::graph::EdgeIndex;
 use itertools::Itertools;
 use std::cmp::{max, min};
 use ahash::RandomState;
+use fxhash::FxHashMap;
 
 use crate::{
     bounds::{bound_exceeded, Bound},
@@ -120,7 +121,7 @@ impl DagNode {
 ///
 /// Return all pairs of non-overlapping, isomorphic subgraphs in the molecule,
 /// sorted to guarantee deterministic iteration.
-pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (HashMap<(usize, usize), usize>, Vec<DagNode>) {
+pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (FxHashMap<(usize, usize), usize>, Vec<DagNode>) {
     let num_edges = mol.graph().edge_count();
 
     let mut subgraphs: Vec<usize> = Vec::new();
@@ -279,7 +280,7 @@ pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (HashMap<(usize, 
 
     // give matches ids
     let mut next_match_id = 0;
-    let mut match_to_id: HashMap<(usize, usize), usize> = HashMap::new();
+    let mut match_to_id: FxHashMap<(usize, usize), usize> = FxHashMap::default();
     for m in matches {
         match_to_id.insert(m, next_match_id);
         next_match_id += 1;
@@ -373,7 +374,7 @@ fn is_usable(state: &[BitSet], h1: &BitSet, h2: &BitSet, masks: &mut Vec<Vec<Bit
 #[allow(clippy::too_many_arguments)]
 pub fn recurse_index_search(
     mol: &Molecule,
-    matches: &HashMap<(usize, usize), usize>,
+    matches: &FxHashMap<(usize, usize), usize>,
     dag: &Vec<DagNode>,
     last_removed: usize,
     mut subgraph: BitSet,
@@ -388,6 +389,12 @@ pub fn recurse_index_search(
 ) -> (usize, usize) {
     //println!("{:?}", removal_order);
     //println!("{:?}", state);
+
+    let mut bucket_time = Duration::new(0, 0);
+    let mut sort_time = Duration::new(0, 0);
+    let mut create_matches_time = Duration::new(0, 0);
+    let mut mask_time = Duration::new(0, 0);
+    let mut bound_time = Duration::new(0, 0);
     
     // Generate matches
     let num_edges = mol.graph().edge_count();
@@ -421,6 +428,7 @@ pub fn recurse_index_search(
                 let child_frag = &dag[*child_id].fragment;
                 let valid = child_frag.is_subset(state_frag);
 
+                let start = Instant::now();
                 if valid {
                     let canon_id = &dag[*child_id].canon_id;
                     
@@ -429,9 +437,11 @@ pub fn recurse_index_search(
                         .and_modify(|bucket| bucket.push((*child_id, state_id)))
                         .or_insert(vec![(*child_id, state_id)]);
                 }
+                bucket_time += start.elapsed();
             }
         }
 
+        let start = Instant::now();
         // Search through buckets and create matches
         for fragments in buckets.values() {
             let mut has_match = BitSet::with_capacity(fragments.len());
@@ -459,7 +469,9 @@ pub fn recurse_index_search(
                 let frag2_state_id = (*frag2.1).1;
 
                 // Check for valid match
-                if let Some(x) = matches.get(&(frag1_id, frag2_id)) {
[…]
