### 2025-09-04 266cdca
Remove itertools.combinations from search
