+++ src/assembly.rs
@@ -389,12 +389,6 @@ pub fn recurse_index_search(
 ) -> (usize, usize) {
     //println!("{:?}", removal_order);
     //println!("{:?}", state);
-
-    let mut bucket_time = Duration::new(0, 0);
-    let mut sort_time = Duration::new(0, 0);
-    let mut create_matches_time = Duration::new(0, 0);
-    let mut mask_time = Duration::new(0, 0);
-    let mut bound_time = Duration::new(0, 0);
     
     // Generate matches
     let num_edges = mol.graph().edge_count();
@@ -428,7 +422,6 @@ pub fn recurse_index_search(
                 let child_frag = &dag[*child_id].fragment;
                 let valid = child_frag.is_subset(state_frag);
 
-                let start = Instant::now();
                 if valid {
                     let canon_id = &dag[*child_id].canon_id;
                     
@@ -437,61 +430,57 @@ pub fn recurse_index_search(
                         .and_modify(|bucket| bucket.push((*child_id, state_id)))
                         .or_insert(vec![(*child_id, state_id)]);
                 }
-                bucket_time += start.elapsed();
             }
         }
 
-        let start = Instant::now();
         // Search through buckets and create matches
         for fragments in buckets.values() {
             let mut has_match = BitSet::with_capacity(fragments.len());
             // Loop over pairs of fragments
-            for pair in fragments.iter().enumerate().combinations(2) {
-                let mut frag1;
-                let mut frag2;
+            for i in 0..fragments.len() {
+                for j in i+1..fragments.len() {
+                    let mut frag1 = (i, fragments[i]);
+                    let mut frag2 = (j, fragments[j]);
 
-                // Order fragments
-                if pair[0].1.0 > pair[1].1.0 {
-                    frag1 = pair[0];
-                    frag2 = pair[1];
-                }
-                else {
-                    frag1 = pair[1];
-                    frag2 = pair[0];
-                }
+                    // Order fragments
+                    if frag1.1.0 < frag2.1.0 {
+                        let temp = frag1;
+                        frag1 = frag2;
+                        frag2 = temp;
+                    }
 
-                let frag1_bucket_idx = frag1.0;
-                let frag1_id = (*frag1.1).0;
-                let frag1_state_id = (*frag1.1).1;
+                    let frag1_bucket_idx = frag1.0;
+                    let frag1_id = frag1.1.0;
+                    let frag1_state_id = frag1.1.1;
 
-                let frag2_bucket_idx = frag2.0;
-                let frag2_id = (*frag2.1).0;
-                let frag2_state_id = (*frag2.1).1;
+                    let frag2_bucket_idx = frag2.0;
+                    let frag2_id = frag2.1.0;
+                    let frag2_state_id = frag2.1.1;
 
-                // Check for valid match
-                // if let Some(x) = matches.get(&(frag1_id, frag2_id)) {
-                if dag[frag1_id].fragment.is_disjoint(&dag[frag2_id].fragment) {
-                    let x = matches.get(&(frag1_id, frag2_id)).unwrap();
-                    // Only add match if it occurs later in the ordering than
-                    // the last removed match
-                    if *x >= last_removed {
-                        valid_matches.push((frag1_id, frag2_id));
+                    // Check for valid match
+                    // if let Some(x) = matches.get(&(frag1_id, frag2_id)) {
+                    if dag[frag1_id].fragment.is_disjoint(&dag[frag2_id].fragment) {
+                        let x = matches.get(&(frag1_id, frag2_id)).unwrap();
+                        // Only add match if it occurs later in the ordering than
+                        // the last removed match
+                        if *x >= last_removed {
+                            valid_matches.push((frag1_id, frag2_id));
 
[…]
