### 2025-09-04 f7b15f3
Move mask creation into match enumeration
