### 2025-09-04 9e7e1f5
Move when memoization happens
