### 2025-09-05 afb46b3
Add simple vector bound
