+++ src/assembly.rs
@@ -32,7 +32,6 @@ use rayon::iter::{IntoParallelRefIterator, ParallelIterator, IndexedParallelIter
 use petgraph::graph::EdgeIndex;
 use itertools::Itertools;
 use std::cmp::{max, min};
-use ahash::RandomState;
 use fxhash::FxHashMap;
 
 use crate::{
@@ -41,7 +40,7 @@ use crate::{
     enumerate::{EnumerateMode},
     kernels::{KernelMode, deletion_kernel, inclusion_kernel},
     memoize::{MemoizeMode, NewCache},
-    molecule::Molecule,
+    molecule::{Bond, Element, Molecule},
     utils::{connected_components_under_edges, edge_neighbors},
     reductions::CompatGraph,
 };
@@ -117,11 +116,17 @@ impl DagNode {
     }
 }
 
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+struct EdgeType {
+    bond: Bond,
+    ends: (Element, Element),
+}
+
 /// Helper function for [`index_search`]; only public for benchmarking.
 ///
 /// Return all pairs of non-overlapping, isomorphic subgraphs in the molecule,
 /// sorted to guarantee deterministic iteration.
-pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (FxHashMap<(usize, usize), usize>, Vec<DagNode>) {
+pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (FxHashMap<(usize, usize), usize>, Vec<DagNode>, Vec<usize>) {
     let num_edges = mol.graph().edge_count();
 
     let mut subgraphs: Vec<usize> = Vec::new();
@@ -286,7 +291,48 @@ pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (FxHashMap<(usize
         next_match_id += 1;
     }
 
-    (match_to_id, dag)
+    // Create edge type ids
+    let graph = mol.graph();
+    let mut edgetype_to_id: HashMap<EdgeType, usize> = HashMap::new();
+    let mut edgetypes: Vec<usize> = vec![0; num_edges];
+    let mut next_id = 0;
+
+    // Create a list of the elements of the nodes
+    // The list can be indexed by NodeIndex-es
+    let mut node_elements: Vec<Element> = Vec::new();
+    for v in graph.node_weights() {
+        node_elements.push(v.element());
+    }
+
+    // Extract bond types
+    let weights: Vec<Bond> = graph.edge_weights().copied().collect();
+
+    // types will hold an element for every unique edge type in fragment
+    for e in graph.edge_indices() {
+        // Extract bond type and endpoint atom types
+        let bond = weights[e.index()];
+        let (end1, end2) = graph.edge_endpoints(e).expect("Edge Endpoint Error");
+
+        let atom1 = node_elements[end1.index()];
+        let atom2 = node_elements[end2.index()];
+
+        // Create edgetype
+        let ends = if atom1 < atom2 {(atom1, atom2)} else {(atom2, atom1)};
+        let edgetype = EdgeType{bond, ends};
+
+        // Find or create id for this edgetype
+        // and store the edgetype id for this edge
+        if let Some(id) = edgetype_to_id.get(&edgetype) {
+            edgetypes[e.index()] = *id;
+        }
+        else {
+            edgetypes[e.index()] = next_id;
+            edgetype_to_id.insert(edgetype, next_id);
+            next_id += 1;
+        }
+    }
+
+    (match_to_id, dag, edgetypes)
 }
 
 /// Determine the fragments produced from the given assembly state by removing
[…]
