### 2025-09-05 e1650d9
Move vector bound
