### 2025-09-09 883444f
Replace bugged vector bound
