+++ Cargo.toml
@@ -18,7 +18,6 @@ petgraph = "0.6.5"
 pyo3 = { version = "0.24.1", features = ["abi3-py310", "extension-module"]}
 rayon = "1.10.0"
 dashmap = "6.1.0"
-ahash = "0.8.12"
 fxhash = "0.2.1"
 itertools = "0.14.0"
 
