### 2025-09-12 535c251
Change vector bound
