### 2025-09-18 d820b0f
New matches struct
