+++ src/assembly.rs
@@ -19,7 +19,7 @@
 //! ```
 
 use std::{
-    collections::{BTreeMap, HashMap, HashSet}, sync::{
+    sync::{
         atomic::{AtomicUsize, Ordering::Relaxed},
         Arc,
     }, time::{Duration, Instant}
@@ -27,22 +27,10 @@ use std::{
 
 use bit_set::BitSet;
 use clap::ValueEnum;
-use graph_canon::canon;
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator, IndexedParallelIterator};
-use petgraph::graph::EdgeIndex;
-use itertools::Itertools;
-use std::cmp::{max, min};
-use fxhash::FxHashMap;
 
 use crate::{
-    bounds::{bound_exceeded, Bound},
-    canonize::{canonize, CanonizeMode, Labeling},
-    enumerate::{EnumerateMode},
-    kernels::{KernelMode, deletion_kernel, inclusion_kernel},
-    memoize::{MemoizeMode, NewCache},
-    molecule::{Bond, Element, Molecule},
-    utils::{connected_components_under_edges, edge_neighbors},
-    reductions::CompatGraph,
+    bounds::Bound, canonize::CanonizeMode, enumerate::EnumerateMode, kernels::KernelMode, matches::Matches, memoize::{MemoizeMode, NewCache}, molecule:: Molecule, utils::{connected_components_under_edges}
 };
 
 /// Parallelization strategy for the recursive search phase.
@@ -99,242 +87,6 @@ pub fn depth(mol: &Molecule) -> u32 {
     ix
 }
 
-#[derive(Debug)]
-pub struct DagNode {
-    fragment: BitSet,
-    canon_id: usize,
-    children: Vec<usize>,
-}
-
-impl DagNode {
-    pub fn new(fragment: BitSet, canon_id: usize) -> Self {
-        Self {
-            fragment,
-            canon_id,
-            children: Vec::new(),
-        }
-    }
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-struct EdgeType {
-    bond: Bond,
-    ends: (Element, Element),
-}
-
-/// Helper function for [`index_search`]; only public for benchmarking.
-///
-/// Return all pairs of non-overlapping, isomorphic subgraphs in the molecule,
-/// sorted to guarantee deterministic iteration.
-pub fn matches(mol: &Molecule, canonize_mode: CanonizeMode) -> (FxHashMap<(usize, usize), usize>, Vec<DagNode>, Vec<usize>) {
-    let num_edges = mol.graph().edge_count();
-
-    let mut subgraphs: Vec<usize> = Vec::new();
-    let mut constructed_frags: HashSet<BitSet> = HashSet::new();
-    let mut frag_to_id: HashMap<BitSet, usize> = HashMap::new();
-
-    let mut dag: Vec<DagNode> = Vec::with_capacity(num_edges);
-    let mut matches: Vec<(usize, usize)> = Vec::new();
-
-    let mut next_canon_id = 0;
-
-    // Generate subgraphs with one edge
-    for i in 0..num_edges {
-        let mut new_bitset = BitSet::with_capacity(num_edges);
-        new_bitset.insert(i);
-        
-        let node = DagNode::new(new_bitset, 0);
-
-        dag.push(node);
-        subgraphs.push(i);
-    }
-
-    next_canon_id += 1;
-
-    while subgraphs.len() > 0 {
-        // Initialize new buckets and subgraph list
-        let mut buckets: BTreeMap<Labeling, Vec<(BitSet, usize)>> = BTreeMap::new();
[…]
