+++ src/matches.rs
@@ -1,0 +1,512 @@
+use std::collections::{BTreeMap, HashMap, HashSet};
+
+use bit_set::BitSet;
+use fxhash::FxHashMap;
+use itertools::Itertools;
+use petgraph::graph::EdgeIndex;
+
+use crate::{bounds::Bound, canonize::{canonize, CanonizeMode, Labeling}, molecule::{Bond, Element, Molecule}, reductions::CompatGraph, utils::{connected_components_under_edges, edge_neighbors}};
+
+pub struct DagNode {
+    fragment: BitSet,
+    canon_id: usize,
+    children: Vec<usize>,
+}
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+struct EdgeType {
+    bond: Bond,
+    ends: (Element, Element),
+}
+
+pub struct Matches {
+    id_to_match: Vec<(usize, usize)>,
+    match_to_id: FxHashMap<(usize, usize), usize>,
+    dag: Vec<DagNode>,
+    // clique: CompatGraph,
+    // clique_offset: usize,
+    edge_types: Vec<usize>,
+}
+
+
+impl DagNode {
+    pub fn new(fragment: BitSet, canon_id: usize) -> Self {
+        Self {
+            fragment,
+            canon_id,
+            children: Vec::new(),
+        }
+    }
+}
+
+impl Matches {
+    // Generate DAG, clique, and other info from the base molecule.
+    pub fn new(mol: &Molecule, canonize_mode: CanonizeMode) -> Self {
+        let num_edges = mol.graph().edge_count();
+
+        let mut subgraphs: Vec<usize> = Vec::new();
+        let mut constructed_frags: HashSet<BitSet> = HashSet::new();
+        let mut frag_to_id: HashMap<BitSet, usize> = HashMap::new();
+
+        let mut dag: Vec<DagNode> = Vec::with_capacity(num_edges);
+        let mut matches: Vec<(usize, usize)> = Vec::new();
+
+        let mut next_canon_id = 0;
+
+        // Generate subgraphs with one edge
+        for i in 0..num_edges {
+            let mut new_bitset = BitSet::with_capacity(num_edges);
+            new_bitset.insert(i);
+            
+            let node = DagNode::new(new_bitset, 0);
+
+            dag.push(node);
+            subgraphs.push(i);
+        }
+
+        next_canon_id += 1;
+
+        while subgraphs.len() > 0 {
+            // Initialize new buckets and subgraph list
+            let mut buckets: BTreeMap<Labeling, Vec<(BitSet, usize)>> = BTreeMap::new();
+            let mut child_subgraphs: Vec<usize> = Vec::new();
+
+            // Extend and bucket new subgraphs
+            for parent_idx in subgraphs.iter() {
+                // Get fragment to be extended
+                let fragment = &dag[*parent_idx].fragment;
+
+                // Construct edge neighborhood of fragment
+                let mut neighborhood = BitSet::with_capacity(num_edges);
+                for e in fragment {
+                    neighborhood.extend(edge_neighbors(mol.graph(), EdgeIndex::new(e)).map(|x| x.index()));
+                }
+                neighborhood.difference_with(fragment);
+
+                // Extend fragment with edges from neighborhood
+                for e in neighborhood.iter() {
+                    // Create extended fragment
+                    let mut new_fragment = fragment.clone();
+                    new_fragment.insert(e);
+
+                    // Check if fragment has already been created
+                    // If yes, continue
+                    if constructed_frags.contains(&new_fragment) {
+                        continue;
+                    }
+
+                    // Bucket subgraph
+                    // Gives ownership of new_fragment to buckets
+                    let canon = canonize(mol, &new_fragment, canonize_mode);
+                    buckets.entry(canon)
[…]
