### 2025-09-18 fd418c0
New state struct
