### 2025-09-18 688bbc6
Refactor bounding
