### 2025-09-18 39b72c4
Build clique
