### 2025-09-18 9a64cad
Fix off by 1 error in state
