### 2025-09-18 eae23de
Make clique an option
