### 2025-09-19 19533da
Add handoff from dag to clique
