### 2025-09-24 e3a7f13
Add new matches and state structs
