+++ src/matches.rs
@@ -1,0 +1,12 @@
+use bit_set::BitSet;
+
+/// TODO
+// This structure will change dramatically in the future
+// so I'm not adding docstrings for now - GP
+pub struct Matches {
+    matches: Vec<(BitSet, BitSet)>,
+}
+
+impl Matches {
+    
+}
