### 2025-09-25 f623945
Move matches generation into matches struct
