### 2025-09-25 ec51b95
Fix how removed match is chosen
