### 2025-09-25 3b702b4
Move state info into struct
