### 2025-09-25 66dfa02
Modify bounds to use state struct
