### 2025-09-25 6862e11
Modify cache to use state struct
