### 2025-09-25 060231f
Use state struct in a few places
