### 2025-09-25 4ce40bb
chore: Fix clippy and fmt lints
