### 2025-09-25 c983baa
docs: Minor cleanup throughout, mostly focused on docstrings
