### 2025-10-01 c10c83c
Remove fragments cache mode and change key function
