### 2025-10-01 1f58e35
Change memoization of state
