### 2025-10-01 191fe63
Add doc strings
