### 2025-10-02 7d293a4
feat: Optimize canonical ID generation using entries, minor polishing
