+++ src/memoize.rs
@@ -32,17 +32,18 @@ pub struct Cache {
     memoize_mode: MemoizeMode,
     /// Canonization mode; only used with [`MemoizeMode::CanonIndex`].
     canonize_mode: CanonizeMode,
-    /// A parallel-aware cache mapping keys (lists of canon ids) to their
-    /// assembly index upper bounds and match removal order.
+    /// A parallel-aware cache mapping keys (lists of usize canonical IDs) to
+    /// their assembly index upper bounds and match removal order.
     cache: Arc<DashMap<Vec<usize>, (usize, Vec<usize>)>>,
-    /// A parallel-aware map from a graph canonization labeling to a canon id.
-    /// These ids are used as keys into the cache since hashing a usize is much
-    /// faster than hashing the labelings directly.
-    label_to_canon_id: Arc<DashMap<Labeling, usize>>,
-    /// A parallel-aware map from fragments (represented as bitsets) to canon
-    /// ids. Two fragments have the same id iff they are isomorphic.
-    frag_to_canon_id: Arc<DashMap<BitSet, usize>>,
-    /// A counter for what id should be given to the next unique labeling seen.
+    /// A parallel-aware map from canonical labelings to canonical IDs. Lists
+    /// of these IDs are used as memoization cache keys since usizes are much
+    /// faster to hash than canonical labelings.
+    labeling_to_id: Arc<DashMap<Labeling, usize>>,
+    /// A parallel-aware map from fragments to canonical IDs. Two fragments
+    /// have the same canonical ID iff they are isomorphic.
+    fragment_to_id: Arc<DashMap<BitSet, usize>>,
+    /// A parallel-aware counter for assigning a unique ID to the next unique
+    /// canonical labeling seen.
     next_id: Arc<AtomicUsize>,
 }
 
@@ -53,24 +54,25 @@ impl Cache {
             memoize_mode,
             canonize_mode,
             cache: Arc::new(DashMap::<Vec<usize>, (usize, Vec<usize>)>::new()),
-            label_to_canon_id: Arc::new(DashMap::<Labeling, usize>::new()),
-            frag_to_canon_id: Arc::new(DashMap::<BitSet, usize>::new()),
+            labeling_to_id: Arc::new(DashMap::<Labeling, usize>::new()),
+            fragment_to_id: Arc::new(DashMap::<BitSet, usize>::new()),
             next_id: Arc::new(AtomicUsize::from(0)),
         }
     }
 
-    /// Create a key into the cache for the given assembly state.
+    /// Create a memoization cache key for the given assembly state.
     ///
-    /// If using [`MemoizeMode::CanonIndex`], keys are sorted lists of canon ids.
+    /// If using [`MemoizeMode::CanonIndex`], keys are sorted lists of
+    /// canonical IDs.
     fn key(&mut self, mol: &Molecule, state: &State) -> Option<Vec<usize>> {
         match self.memoize_mode {
             MemoizeMode::None => None,
             MemoizeMode::CanonIndex => {
-                let mut frag_ids = Vec::new();
-                for frag in state.fragments() {
-                    let id = self.get_canon_id(mol, frag);
-                    frag_ids.push(id);
-                }
+                let mut frag_ids: Vec<usize> = state
+                    .fragments()
+                    .iter()
+                    .map(|fragment| self.canonical_id(mol, fragment))
+                    .collect();
                 frag_ids.sort();
 
                 Some(frag_ids)
@@ -78,30 +80,15 @@ impl Cache {
         }
     }
 
-    /// Takes a fragment and canonizes it using the specified [`CanonizeMode`],
-    /// then returns the corresponding canon id.
[…]
