### 2025-10-02 eb9011b
fix: Remove MemoizeMode::FragsIndex from tests and benchmarks
