### 2025-10-02 8c5bd8a
chore: Ignore specific clippy lint about complex data type
