### 2025-10-02 0b1b225
feat: Restore non-entry version of canonical ID generation
