### 2025-10-13 085b208
Add DagNode struct
