+++ src/matches.rs
@@ -1,7 +1,10 @@
 //! Pairs of non-overlapping, isomorphic subgraphs in a molecule.
 
+use std::collections::{BTreeMap, HashMap, HashSet};
+
 use bit_set::BitSet;
 use dashmap::DashMap;
+use petgraph::graph::EdgeIndex;
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 
 use crate::{
@@ -10,6 +13,7 @@ use crate::{
     enumerate::{enumerate_subgraphs, EnumerateMode},
     molecule::Molecule,
     state::State,
+    utils::edge_neighbors,
 };
 
 struct DagNode {
@@ -21,7 +25,9 @@ struct DagNode {
 /// Pairs of non-overlapping, isomorphic subgraphs in a molecule, sorted to
 /// guarantee deterministic iteration.
 pub struct Matches {
-    matches: Vec<(BitSet, BitSet)>,
+    dag: Vec<DagNode>,
+    id_to_match: Vec<(usize, usize)>,
+    match_to_id: HashMap<(usize, usize), usize>,
 }
 
 impl DagNode {
@@ -50,72 +56,195 @@ impl Matches {
         canonize_mode: CanonizeMode,
         parallel_mode: ParallelMode,
     ) -> Self {
-        // Enumerate all connected, non-induced subgraphs with at most |E|/2
-        // edges and bin them into isomorphism classes using canonization.
-        let isomorphism_classes = DashMap::<Labeling, Vec<BitSet>>::new();
-        let bin_subgraph = |subgraph: &BitSet| {
-            isomorphism_classes
-                .entry(canonize(mol, subgraph, canonize_mode))
-                .and_modify(|bucket| bucket.push(subgraph.clone()))
-                .or_insert(vec![subgraph.clone()]);
-        };
-        if parallel_mode == ParallelMode::None {
-            enumerate_subgraphs(mol, enumerate_mode)
-                .iter()
-                .for_each(bin_subgraph);
-        } else {
-            enumerate_subgraphs(mol, enumerate_mode)
-                .par_iter()
-                .for_each(bin_subgraph);
+        let num_edges = mol.graph().edge_count();
+
+        let mut subgraph_ids: Vec<usize> = Vec::new();
+        let mut constructed_frags: HashSet<BitSet> = HashSet::new();
+        let mut dag_ids: HashMap<BitSet, usize> = HashMap::new();
+        let mut dag: Vec<DagNode> = Vec::with_capacity(num_edges);
+        let mut matches: Vec<(usize, usize)> = Vec::new();
+        let mut next_canonical_id = 1;
+
+        // Helper functions
+        fn add_to_dag(
+            dag: &mut Vec<DagNode>,
+            fragment: &BitSet,
+            parent_id: usize,
+            canonical_id: usize,
+        ) -> usize {
+            // Create new dag node
+            let new_dag_node = DagNode::new(fragment.clone(), canonical_id);
+            let child_id = dag.len();
+
+            // Add child to dag list
+            dag.push(new_dag_node);
+
+            // Add index to parent's child list
+            let parent = &mut dag[parent_id];
+            parent.children.push(child_id);
+
+            child_id
         }
 
-        // In each isomorphism class, get non-overlapping pairs of subgraphs.
-        let mut matches = Vec::new();
-        for bucket in isomorphism_classes.iter() {
-            for (i, first) in bucket.iter().enumerate() {
-                for second in &bucket[i..] {
-                    if first.is_disjoint(second) {
-                        if first > second {
-                            matches.push((first.clone(), second.clone()));
-                        } else {
[…]
