### 2025-10-14 ab1cc28
Change assembly recursion to use match ids
