### 2025-10-14 c9138e6
Remove enumeration
