### 2025-10-15 b350554
Doc strings and clippy lints
