### 2025-10-15 c38feb6
fix: Remove missed enumeration content from Python and doc tests
