### 2025-10-17 2b7908f
refactor: Significant commenting and cleanup of new matches code
