+++ src/matches.rs
@@ -1,5 +1,5 @@
-//! Strucutral information on matches in the molecular graph - pairs of
-//! non-overlapping, isomorphic subgraphs
+//! Strucutral information on "matches" in a molecular graph, i.e., pairs of
+//! edge-disjoint, isomorphic subgraphs.
 
 use std::collections::{BTreeMap, HashMap, HashSet};
 
@@ -13,35 +13,38 @@ use crate::{
     utils::edge_neighbors,
 };
 
-/// Holds data for nodes of the dag
+/// A node in the DAG storing fragment information; see [`Matches`].
 struct DagNode {
-    /// Connected subgraph of the molecule
+    /// The fragment (i.e., connected molecular subgraph) this node represents.
     fragment: BitSet,
-    /// IDs for isomorphism. Two DagNodes have the same canonical_id iff their
-    /// fragments are ismorphic.
+    /// The canonical ID of this node's fragment. Two [`DagNode`]s have the
+    /// same canonical ID iff their fragments are isomorphic.
     canonical_id: usize,
-    /// List of indices of this nodes children. If A is a child of B then
-    /// A.fragment is B.fragment with an additional edge.
+    /// Indices of this node's children/out-neighbors in the DAG. If u is a
+    /// child of v then u.fragment is v.fragment with an additional edge.
     children: Vec<usize>,
 }
 
-/// Structural information on the molecules matches - pairs of nonoverlapping
-/// isomorphic subgraph.
+/// Structural information on "matches" in a molecular graph, i.e., pairs of
+/// edge-disjoint, isomorphic subgraphs.
 pub struct Matches {
-    /// Each node of the dag holds information on a duplicatable subgraph (i.e.
-    /// subgraphs such that there exists a disjoint ismorphic copy of the
-    /// subgraph in the molecule).
+    /// Seet et al. (2024) perform match enumeration by constructing a directed
+    /// acyclic graph (DAG). Each node in this DAG is a fragment (i.e., a
+    /// connected molecular subgraph) that is duplicatable (i.e., there exists
+    /// some other edge-disjoint fragment it is isomorphic to). If there is an
+    /// edge from u to v, then fragment v is fragment u with one added edge
+    /// (note: this new edge may be between two existing nodes in u or "extend"
+    /// from one existing node in u to a new node).
     dag: Vec<DagNode>,
-    /// List of all possible matches stored as a pair of fragment IDs (the
-    /// index of the fragment in the dag).
-    id_to_match: Vec<(usize, usize)>,
-    /// Given a match (as a pair of fragment IDs) returns its match ID, i.e.,
-    /// its index in id_to_match
-    match_to_id: HashMap<(usize, usize), usize>,
+    /// All possible matches (i.e., pairs of edge-disjoint, isomorphic
+    /// fragments) stored as pairs of fragment (i.e., DAG node) indices.
+    matches: Vec<(usize, usize)>,
+    /// Maps pairs of matched fragment indices to their indices in `matches`.
+    match_to_ix: HashMap<(usize, usize), usize>,
 }
 
 impl DagNode {
-    /// Create new [`DagNode`] object.
+    /// Create a new [`DagNode`].
     pub fn new(fragment: BitSet, canonical_id: usize) -> Self {
         Self {
             fragment,
@@ -52,282 +55,220 @@ impl DagNode {
 }
 
 impl Matches {
-    /// Generate [`Matches`] from the given molecule with the specified modes.
+    /// Generate [`Matches`] from the given molecule and canonization mode.
     pub fn new(mol: &Molecule, canonize_mode: CanonizeMode) -> Self {
         let num_edges = mol.graph().edge_count();
-
[…]
