### 2025-10-18 057a85a
docs: Refer to matches as "edge-disjoint" instead of "non-overlapping"
