+++ src/assembly.rs
@@ -93,7 +93,7 @@ pub fn depth(mol: &Molecule) -> u32 {
 }
 
 /// Determine the fragments produced from the given assembly state by removing
-/// the given pair of non-overlapping, isomorphic subgraphs and then adding one
+/// the given pair of edge-disjoint, isomorphic subgraphs and then adding one
 /// back; return `None` if not possible.
 fn fragments(mol: &Molecule, state: &[BitSet], h1: &BitSet, h2: &BitSet) -> Option<Vec<BitSet>> {
     // Attempt to find fragments f1 and f2 containing h1 and h2, respectively;
@@ -143,7 +143,7 @@ fn fragments(mol: &Molecule, state: &[BitSet], h1: &BitSet, h2: &BitSet) -> Opti
 ///
 /// Inputs:
 /// - `mol`: The molecule whose assembly index is being calculated.
-/// - `matches`: The non-overlapping isomorphic subgraph pairs left to check.
+/// - `matches`: Structural information about the molecule's matched fragments.
 /// - `state`: The current assembly state.
 /// - `best_index`: The smallest assembly index for all assembly states so far.
 /// - `bounds`: The list of bounding strategies to apply.
@@ -172,7 +172,8 @@ pub fn recurse_index_search(
         return (state.index(), 1);
     }
 
-    // Generate list of matches to try removing from this state.
+    // Generate a list of matches (i.e., pairs of edge-disjoint, isomorphic
+    // fragments) to remove from this state.
     let matches_to_remove = matches.matches_to_remove(state);
 
     // Keep track of the best assembly index found in any of this assembly
@@ -181,7 +182,7 @@ pub fn recurse_index_search(
     let states_searched = AtomicUsize::from(1);
 
     // Define a closure that handles recursing to a new assembly state based on
-    // the given match (i.e., pair of non-overlapping isomorphic subgraphs).
+    // the given match.
     let recurse_on_match = |i: usize, match_ix: usize| {
         let (h1, h2) = matches.match_fragments(match_ix);
 
@@ -242,7 +243,7 @@ pub fn recurse_index_search(
 ///
 /// The results returned are:
 /// - The molecule's `u32` assembly index.
-/// - The molecule's `u32` number of non-overlapping isomorphic subgraph pairs.
+/// - The molecule's `u32` number of edge-disjoint isomorphic subgraph pairs.
 /// - The `usize` total number of assembly states searched, where an assembly
 ///   state is a collection of fragments. Note that, depending on the algorithm
 ///   parameters used, some states may be searched/counted multiple times.
@@ -305,7 +306,7 @@ pub fn index_search(
         panic!("The chosen --kernel mode is not implemented yet!")
     }
 
-    // Enumerate non-overlapping isomorphic subgraph pairs.
+    // Enumerate matches (i.e., pairs of edge-disjoint isomorphic fragments).
     let matches = Matches::new(mol, canonize_mode);
 
     // Create memoization cache.
