### 2025-10-24 36f73f4
Compute used edge masks
