### 2025-11-02 DaymudeLab/assembly-theory#123 @amcandio
[JOSS review] Improve discoverability of python installation instructions
Hi! For a minute I thought python installation instruction were missing. I'd suggest you improve discoverability by mentioning them in the main README file

https://github.com/openjournals/joss-reviews/issues/9318
