### 2025-11-02 DaymudeLab/assembly-theory#124 @amcandio
[JOSS review] Examples opening files
Current examples are great but I believe they could be improved by removing the friction of opening a file. For example, right now they do:

```python
import assembly_theory as at  # Note the '_' instead of '-'!

# Load a mol block from file.
with open('anthracene.mol') as f:
    mol_block = f.read()

# Calculate the molecule's assembly index.
at.index(mol_block)  # 6
```

But `anthracene.mol` is not a huge file, so you could replace it with the whole content in a string variable. You can also use a shorter file such as `benzene.mol` for this. This way, user can just copy and paste to terminal to try it out.

https://github.com/openjournals/joss-reviews/issues/9318
