### 2025-11-02 DaymudeLab/assembly-theory#125 @amcandio
[JOSS review] Functionality documentation
The key function [index](https://docs.rs/assembly-theory/latest/assembly_theory/assembly/fn.index.html) should point to how to tune the default strategy for more advanced users. Also, it is unclear what python entry points are available and what parameters they take.

https://github.com/openjournals/joss-reviews/issues/9318
