### 2025-11-15 DaymudeLab/assembly-theory#126 @amcandio
[JOSS review] Missing guidelines on how to run python tests
I had to manually install `pytest`. It would help to have better guidance on installing dependencies to test python library locally.

https://github.com/openjournals/joss-reviews/issues/9318
