### 2025-11-15 DaymudeLab/assembly-theory#127 @amcandio
[JOSS review] Manuscript feedback
Hi! I just have some minor comments on the manuscript

> Previous software to compute MA have been approximate, closed-source, platform-dependent, or written in languages rarely used by the broader scientific community.

It would be great to mention and cite (if applicable) which other software that is.

> The more recent assembly_go implementation computes MA exactly but is written in Go, yielding worse performance and posing an accessibility barrier for most scientists who are unfamiliar with the language

This performance claim is not substaint by any benchmarking results. Regarding language familiarity, "barrier for most scientist" can be deemed subjective. I'd suggest to instead focus on AT library being integrated with Python (objectively popular science language) as opposed to saying "most scientist have a barrier with go" (which is harder to prove).

https://github.com/openjournals/joss-reviews/issues/9318
