### 2025-11-17 DaymudeLab/assembly-theory#123 @abhishektiwari
+1 to this. Add following to README.md of the repository,

```
pip install assembly-theory   # Using pip.
pipx install assembly-theory  # Using pipx.
uv add assembly-theory        # Using uv.
```
