### 2025-11-17 DaymudeLab/assembly-theory#127 @abhishektiwari
+1. Unclear if performance is better than other implementations as claimed in paper. Following claim is not supported by data.

> extensible, high-performance Rust implementations of assembly index calculation algorithms

I know there is a benchmark script but how results compare with Go side-by-side.
