### 2025-11-18 3a00159
docs: Add Rust crate/Python package install instructions to main README
