### 2025-11-18 07aaae0
docs: Point from index to index_search if users want customization
