### 2025-11-18 69d48ca
docs: Add explicit pytest installation commands to README
