### 2025-11-18 be38bfe
docs: Include anthracene mol block directly in Python usage examples
