### 2025-11-24 e0b6ce8
docs: Give Python README a dedicated API Reference header
