### 2025-11-24 4a89357
Fix state update bug
