### 2025-11-25 1824ea5
New match generation and bound
