+++ src/assembly.rs
@@ -28,7 +28,7 @@ use clap::ValueEnum;
 use rayon::iter::{IndexedParallelIterator, IntoParallelRefIterator, ParallelIterator};
 
 use crate::{
-    bounds::{bound_exceeded, Bound},
+    bounds::Bound,
     canonize::CanonizeMode,
     kernels::KernelMode,
     matches::Matches,
@@ -166,15 +166,15 @@ pub fn recurse_index_search(
 ) -> (usize, usize) {
     // If any bounds would prune this assembly state or if memoization is
     // enabled and this assembly state is preempted by the cached state, halt.
-    if bound_exceeded(mol, state, best_index.load(Relaxed), bounds)
-        || cache.memoize_state(mol, state)
+    if crate::bounds::bound_exceeded(mol, state, best_index.load(Relaxed), bounds)
+       || cache.memoize_state(mol, state) 
     {
         return (state.index(), 1);
     }
 
     // Generate a list of matches (i.e., pairs of edge-disjoint, isomorphic
     // fragments) to remove from this state.
-    let matches_to_remove = matches.matches_to_remove(mol, state);
+    let (intermediate_frags, matches_to_remove): (Vec<BitSet>, Vec<usize>) = matches.matches_to_remove(mol, state, best_index.load(Relaxed), bounds);
 
     // Keep track of the best assembly index found in any of this assembly
     // state's children and the number of states searched, including this one.
@@ -186,7 +186,7 @@ pub fn recurse_index_search(
     let recurse_on_match = |i: usize, match_ix: usize| {
         let (h1, h2) = matches.match_fragments(match_ix);
 
-        if let Some(fragments) = fragments(mol, state.fragments(), h1, h2) {
+        if let Some(fragments) = fragments(mol, &intermediate_frags, h1, h2) {
             // If using depth-one parallelism, all descendant states should be
             // computed serially.
             let new_parallel = if parallel_mode == ParallelMode::DepthOne {
@@ -199,7 +199,7 @@ pub fn recurse_index_search(
             let (child_index, child_states_searched) = recurse_index_search(
                 mol,
                 matches,
-                &state.update(fragments, i, h1.len()),
+                &state.update(fragments, i, match_ix, h1.len()),
                 best_index.clone(),
                 bounds,
                 &mut cache.clone(),
