+++ src/matches.rs
@@ -185,12 +185,17 @@ impl Matches {
         self.matches.is_empty()
     }
 
-    /// Return all removable matches for the given assembly state.
+    /// Return all removable matches that may result in a better assembly index
+    /// for the given assembly state.
     ///
     /// A match (i.e., two edge-disjoint isomorphic fragments) is removable
     /// from an assembly state if (1) each match fragment is a subgraph of some
     /// assembly state fragment, and (2) the match's index is strictly greater
     /// than that of the last match removed from this assembly state.
+    /// 
+    /// Bounds are used to find matches whose removal will not lead to a better
+    /// assembly index that what has already been found. These matches will not
+    /// be included in the returned list.
     pub fn matches_to_remove(&self, mol: &Molecule, state: &State, best: usize, bounds: &[Bound]) -> (Vec<BitSet>, Vec<usize>) {
         // The search for removable matches uses the DAG of duplicatable
         // fragments, starting with singleton edge fragments (the DAG's source
@@ -202,10 +207,16 @@ impl Matches {
             frag_state_ixs.extend(frag.iter().map(|e| (e, state_ix)));
         }
 
-        let mut masks: Vec<Vec<BitSet>> = Vec::new();
-        let mut buckets_by_len: Vec<BTreeMap<usize, Vec<(usize, usize)>>> = Vec::new();
+        // matchable_edge_masks has an element for each size of a removable
+        // match in this assembly state (the size of a match is the number of
+        // edges in its fragments). For a given size, stores the edges of each
+        // state fragment that are included in some match of this size. This
+        // information is used for later bounding.
+        let mut matchable_edge_masks: Vec<Vec<BitSet>> = Vec::new();
+        let mut iso_classes_by_len: Vec<BTreeMap<usize, Vec<(usize, usize)>>> = Vec::new();
 
-        // Search for removable matches using BFS over the DAG, only extending
+        // Collect matchable fragments by isomorphism and find which edges are
+        // used in a removable match. Use BFS over the DAG, only extending
         // search paths that include fragments contained in removable matches.
         // (If some duplicatable fragment is not in a removable match, then no
         // child fragment that extends it with additional edges can be either.)
@@ -235,15 +246,17 @@ impl Matches {
             // Iterate through each isomorphism class (containing isomorphic
             // removable fragments) to identify removable matches.
             let mut next_frag_state_ixs: Vec<(usize, usize)> = Vec::new();
-            let mut used_edge_mask =
+            let mut matchable_edges =
                 vec![BitSet::with_capacity(mol.graph().edge_count()); state.fragments().len()];
             for isomorphic_frags in isomorphism_classes.values_mut() {
                 // Track which fragments in this isomorphism class are matched.
                 let mut frag_has_match = BitSet::with_capacity(isomorphic_frags.len());
 
-                // Check if pairs of isomorphic fragments form a match.
+                // Check for which fragments are used in a removable match.
                 for iso_ix1 in 0..isomorphic_frags.len() {
                     for iso_ix2 in (iso_ix1 + 1)..isomorphic_frags.len() {
+                        // If both fragments have already been found to be
+                        // matchable, continue.
                         if frag_has_match.contains(iso_ix1) && frag_has_match.contains(iso_ix2) {
                             continue;
                         }
@@ -268,12 +281,12 @@ impl Matches {
                                 // these fragments on first match.
                                 if frag_has_match.insert(frag1_iso_ix) {
                                     next_frag_state_ixs.push((frag1_ix, frag1_state_ix));
-                                    used_edge_mask[frag1_state_ix]
+                                    matchable_edges[frag1_state_ix]
                                         .union_with(&self.dag[frag1_ix].fragment);
                                 }
[…]
