### 2025-11-25 b9edf67
Renaming and docstrings
