### 2025-11-25 536a17f
Make usable-edges a default bound
