### 2025-11-26 d2edcd0
Cargo fmt and clippy lints
