### 2025-11-26 59035af
Change parallel iterator of index search to par_bridge
