### 2025-11-30 e1a23a8
Creating AssemblyObject traits
