### 2025-11-30 6e577c6
Revert "Creating AssemblyObject traits"
This reverts commit e1a23a8a13b088ca52fa0b61228fc563bc190555.
