### 2025-11-30 81f4b4e
stubbing out Traits
