### 2025-12-01 52ea1b8
Working example with new Trait based index calculations
