### 2025-12-01 a08e6f0
Actually now working on Trait based implementation, playing with timing to check for significant regression
