### 2025-12-01 2f1fff4
Trying to expose string assembly to python
