### 2025-12-03 314599a
docs: Polish bounds and matches documentation, state vs. match bounds
Also, restrict visibility of bound functions to crate public
