### 2025-12-03 c772b1d
refactor: Remove unimplemented bounds, set new default bounds
