+++ src/main.rs
@@ -90,12 +90,7 @@ fn main() -> Result<()> {
     // Handle bounding strategy CLI arguments.
     let boundlist: &[Bound] = match cli.boundsgroup {
         // By default, use a combination of the integer and vector bounds.
-        None => &[
-            Bound::Int,
-            Bound::VecSimple,
-            Bound::VecSmallFrags,
-            Bound::UsableEdges,
-        ],
+        None => &[Bound::Int, Bound::MatchableEdges],
         // If --no-bounds is set, do not use any bounds.
         Some(BoundsGroup {
             no_bounds: true, ..
