### 2025-12-03 558724b
fix: Update tests, benchmarks, and Python interface with new bounds
