+++ python/tests/test_assembly_theory.py
@@ -70,7 +70,7 @@ def test_index_search():
             "none",  # Disable parallelism for deterministic states_searched.
             "none",
             "none",
-            ["int", "vec-simple", "vec-small-frags"])
+            ["int", "matchable-edges"])
 
     assert (index, num_matches, states_searched) == (6, 466, 2462)
 
