### 2025-12-03 2e6cfb1
fix: Update Python tests with new number of states searched
