### 2025-12-04 51ab06c
Update strings to handle independent strings (joint assembly support)
