### 2025-12-04 251e5d4
paper: Add a README for paper PDF/benchmarks and benchmark submodules
