### 2025-12-04 99d692f
paper: Modernize JOSS assembly-theory benchmark
