### 2025-12-04 32050f3
paper: Modernize JOSS assembly_go benchmark
