### 2025-12-04 4252aa2
paper: Move figures back into their own directory
