+++ paper/paper.md
@@ -68,7 +68,7 @@ Machine learning methods only provide approximate MA values [@Gebhard2022-inferr
 The more recent `assembly_go` implementation computes MA exactly but is written in Go, yielding worse performance and posing an accessibility barrier for most scientists who are unfamiliar with the language [@Jirasek2024-investigatingquantifying].
 Finally, the latest `assemblycpp` C++ implementation achieves significant performance milestones through a branch-and-bound approach, but until very recently was not publicly available for comparison or verification by the community [@Seet2024-rapidcomputation].
 
-![*Assembly Pathways for Anthracene*. Starting with bonds as building blocks (yellow), a joining operation yields progressively larger structures by combining any two compatible structures that have already been constructed (arrows). These intermediate structures must obey valence rules but otherwise do not have to be physically accessible or chemically synthesizable. There may be many assembly pathways from building blocks to a target structure&mdash;in this case, Anthracene (green)&mdash;but the length of any shortest such pathway (blue) is that structure's assembly index.\label{fig:assemblyindex}](anthracene.pdf){ width=100% }
+![*Assembly Pathways for Anthracene*. Starting with bonds as building blocks (yellow), a joining operation yields progressively larger structures by combining any two compatible structures that have already been constructed (arrows). These intermediate structures must obey valence rules but otherwise do not have to be physically accessible or chemically synthesizable. There may be many assembly pathways from building blocks to a target structure&mdash;in this case, Anthracene (green)&mdash;but the length of any shortest such pathway (blue) is that structure's assembly index.\label{fig:assemblyindex}](figures/anthracene.pdf){ width=100% }
 
 With `assembly-theory`, we provide an open-source, fully documented, extensible, and high-performance library for assembly index calculation, implementing not just one algorithm, but a framework and source of ground truth within which current and future algorithmic approaches can be validated and compared.
 The main implementation is written in Rust, which we chose for its cross-platform support, memory-safety, performant runtime, convenient parallelism, and integrated testing and documentation [@Perkel2020-whyscientists].
