### 2025-12-04 5e615d8
paper: Strip JOSS assembly-theory benchmark to the default strategy
