+++ paper/scripts/benchmark.rs
@@ -1,18 +1,10 @@
 use std::{ffi::OsStr, fs, path::Path};
 
-use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
+use criterion::{criterion_group, criterion_main, Criterion};
 
-use assembly_theory::{
-    assembly::{index_search, ParallelMode},
-    bounds::Bound,
-    canonize::CanonizeMode,
-    kernels::KernelMode,
-    loader::parse_molfile_str,
-    memoize::MemoizeMode,
-    molecule::Molecule,
-};
+use assembly_theory::{assembly::index, loader::parse_molfile_str, molecule::Molecule};
 
-/// Parse all .mol files in `dataset` as [`Molecule`]s.
+/// Parse all .mol files in the given dataset as [`Molecule`]s.
 fn load_dataset_molecules(dataset: &str) -> Vec<Molecule> {
     let paths = fs::read_dir(Path::new("data").join(dataset)).unwrap();
     let mut mol_list: Vec<Molecule> = Vec::new();
@@ -31,43 +23,21 @@ fn load_dataset_molecules(dataset: &str) -> Vec<Molecule> {
     mol_list
 }
 
-/// Benchmark the entire [`index_search`] function for different [`Bound`]s.
+/// Benchmark the [`index`] function for different reference datasets.
 pub fn joss_bench(c: &mut Criterion) {
     // Define a new criterion benchmark group for the JOSS manuscript.
     let mut bench_group = c.benchmark_group("joss");
 
-    // Define datasets and bound lists.
-    let datasets = ["gdb13_1201", "gdb17_200", "checks", "coconut_55"];
-    let bound_lists = [
-        (vec![], "no-bounds"),
-        (vec![Bound::Log], "log"),
-        (vec![Bound::Int], "int"),
-        (vec![Bound::Int, Bound::MatchableEdges], "int-matchable"),
-    ];
-
-    // Run the benchmark for each dataset and bound list.
-    for dataset in &datasets {
+    // Run the benchmark for each reference dataset.
+    for dataset in ["gdb13_1201", "gdb17_200", "checks", "coconut_55"] {
         let mol_list = load_dataset_molecules(dataset);
-        for (bounds, name) in &bound_lists {
-            bench_group.bench_with_input(
-                BenchmarkId::new(*dataset, &name),
-                &bounds,
-                |b, &bounds| {
-                    b.iter(|| {
-                        for mol in &mol_list {
-                            index_search(
-                                &mol,
-                                CanonizeMode::TreeNauty,
-                                ParallelMode::DepthOne,
-                                MemoizeMode::CanonIndex,
-                                KernelMode::None,
-                                &bounds,
-                            );
-                        }
-                    });
-                },
-            );
-        }
+        bench_group.bench_function(dataset, |b| {
+            b.iter(|| {
+                for mol in &mol_list {
+                    index(&mol);
+                }
+            });
+        });
     }
 
     bench_group.finish();
