### 2025-12-04 28a2ef3
paper: Rebuild JOSS uv environment and benchmark statistics script
