+++ paper/pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "joss-scripts"
+version = "0.1.0"
+description = "Data parsing scripts for the assembly-theory JOSS manuscript"
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "numpy>=2.3.5",
+    "pandas>=2.3.3",
+    "scipy>=1.16.3",
+]
