### 2025-12-05 2b7eccb
paper: First attempt at a JOSS assemblycpp-v5 benchmark, but it's slow?
