### 2025-12-05 6c111e9
paper: Complete JOSS assemblycpp-v5 benchmark
