### 2025-12-05 2f3afc3
paper: Rename paper/scripts to paper/benches
