### 2025-12-05 3bd568d
paper: Add performance evaluation to the paper
