+++ paper/paper.md
@@ -65,14 +65,15 @@ In general, exact MA calculation is NP-hard [@Kempes2024-assemblytheory].
 Previous software to compute MA have been approximate, closed-source, platform-dependent, or written in languages rarely used by the broader scientific community.
 The original software to compute a split-branch approximation of MA (an upper bound on the exact value) was written in C++ and depended on the MSVC compiler, making it difficult to deploy to non-Windows machines [@Marshall2021-identifyingmolecules].
 Machine learning methods only provide approximate MA values [@Gebhard2022-inferringmolecular].
-The more recent `assembly_go` implementation computes MA exactly but is written in Go, yielding worse performance and posing an accessibility barrier for most scientists who are unfamiliar with the language [@Jirasek2024-investigatingquantifying].
-Finally, the latest `assemblycpp` C++ implementation achieves significant performance milestones through a branch-and-bound approach, but until very recently was not publicly available for comparison or verification by the community [@Seet2024-rapidcomputation].
+The more recent `assembly_go` computes MA exactly but is written in Go and implements a somewhat naive algorithm, yielding prohibitively slow performance even on mid-size molecules [@Jirasek2024-investigatingquantifying].
+Finally, the latest `assemblycpp-v5` C++ implementation achieves significant performance milestones through an improved branch-and-bound approach, but lacks parallelism, has significant readability and maintenance barriers, and until recently was not publicly available for comparison or verification by the community [@Seet2024-rapidcomputation].
 
 ![*Assembly Pathways for Anthracene*. Starting with bonds as building blocks (yellow), a joining operation yields progressively larger structures by combining any two compatible structures that have already been constructed (arrows). These intermediate structures must obey valence rules but otherwise do not have to be physically accessible or chemically synthesizable. There may be many assembly pathways from building blocks to a target structure&mdash;in this case, Anthracene (green)&mdash;but the length of any shortest such pathway (blue) is that structure's assembly index.\label{fig:assemblyindex}](figures/anthracene.pdf){ width=100% }
 
-With `assembly-theory`, we provide an open-source, fully documented, extensible, and high-performance library for assembly index calculation, implementing not just one algorithm, but a framework and source of ground truth within which current and future algorithmic approaches can be validated and compared.
+With `assembly-theory`, we provide an open-source, fully documented, extensible, and high-performance library for assembly index calculation.
+It moves beyond an implementation of a single algorithm, instead acting a framework and source of ground truth within which current and future algorithmic approaches can be validated and compared.
 The main implementation is written in Rust, which we chose for its cross-platform support, memory-safety, performant runtime, convenient parallelism, and integrated testing and documentation [@Perkel2020-whyscientists].
-We also provide Python bindings, enabling ease of use for scientific practitioners.
+We also leverage modern Rust tooling to provide native Python bindings, enabling ease of use for scientific practitioners and integration with existing Python cheminformatics libraries.
 
 
 
@@ -168,10 +169,36 @@ These reference data are sampled from:
 - COCONUT, a database of natural products (secondary metabolites) offering a rich source of evolved chemical complexity [@Sorokina2021-coconutonline; @Chandrasekhar2025-coconut20].
 
[…]
