### 2025-12-05 86c9f50
docs: Minor README fixes
