### 2025-12-08 ec7329b
paper: Update Seet reference to in-press journal version
