+++ paper/paper.bib
@@ -345,15 +345,12 @@
   doi = {10.1021/ci300415d}
 }
 
-@misc{Seet2024-rapidcomputation,
-  title = {Rapid {{Computation}} of the {{Assembly Index}} of {{Molecular Graphs}}},
+@article{Seet2025-rapidexploration,
+  title = {Rapid {{Exploration}} of the {{Assembly Chemical Space}} of {{Molecular Graphs}}},
   author = {Seet, Ian and Patarroyo, Keith Y. and Siebert, Gage and Walker, Sara I. and Cronin, Leroy},
-  number = {2410.09100},
-  year = {2024},
-  publisher = {arXiv},
-  doi = {10.48550/arXiv.2410.09100},
-  eprint = {2410.09100},
-  archiveprefix = {arXiv}
+  journal = {Journal of Chemical Information and Modeling},
+  year = {2025},
+  doi = {10.1021/acs.jcim.5c01964}
 }
 
 @article{Sharma2023-assemblytheory,
