### 2025-12-08 e01db6e
docs: Update Seet et al. (2025) citation throughout
