### 2025-12-08 6abf702
docs: Update Seet et al. (2025) citation throughout
