### 2025-12-09 f4d271b
docs: Add JOSS shield to README
