### 2025-12-10 d17c6b9
build: Add tokio as a dependency for convenient async and timeouts
