### 2025-12-10 e2e9aff
feat: First working version of timeouts across CLI, Rust, and Python
