### 2025-12-10 d24c3ba
refactor: Rename timeout_ms as timeout to align with CLI and Python
