### 2025-12-10 691912b
fix: Fix broken index_search tests
