### 2025-12-11 25c9ad2
refactor: Make states_searched an Arc shared across recursive calls
