### 2025-12-11 53b71f2
feat: Indicate whether assembly::index_search timed out in new return
