+++ tests/reference_datasets.rs
@@ -71,7 +71,7 @@ fn test_reference_dataset(
         .expect(&format!("Failed to parse {name:?}"));
 
         // Calculate the molecule's assembly index.
-        let (index, _, _) = index_search(
+        let (index, _, _, _) = index_search(
             &mol,
             None,
             canonize_mode,
