+++ python/tests/test_assembly_theory.py
@@ -13,14 +13,14 @@ import assembly_theory as at
 
 
 def test_mol_info():
-    with open(osp.join('data', 'checks', 'anthracene.mol')) as f:
+    with open(osp.join("data", "checks", "anthracene.mol")) as f:
         mol_block = f.read()
 
     info = at.mol_info(mol_block)
-    num_atoms = info.count("label = \"Atom")
-    num_single = info.count("label = \"Single\"")
-    num_double = info.count("label = \"Double\"")
-    num_triple = info.count("label = \"Triple\"")
+    num_atoms = info.count('label = "Atom')
+    num_single = info.count('label = "Single"')
+    num_double = info.count('label = "Double"')
+    num_triple = info.count('label = "Triple"')
 
     assert (num_atoms, num_single, num_double, num_triple) == (14, 9, 7, 0)
 
@@ -33,7 +33,7 @@ def test_mol_info_bad_molblock():
 
 
 def test_depth():
-    with open(osp.join('data', 'checks', 'benzene.mol')) as f:
+    with open(osp.join("data", "checks", "benzene.mol")) as f:
         mol_block = f.read()
 
     assert at.depth(mol_block) == 3
@@ -47,7 +47,7 @@ def test_depth_bad_molblock():
 
 
 def test_index():
-    with open(osp.join('data', 'checks', 'anthracene.mol')) as f:
+    with open(osp.join("data", "checks", "anthracene.mol")) as f:
         mol_block = f.read()
 
     assert at.index(mol_block) == 6
@@ -61,19 +61,18 @@ def test_index_bad_molblock():
 
 
 def test_index_search():
-    with open(osp.join('data', 'checks', 'anthracene.mol')) as f:
+    with open(osp.join("data", "checks", "anthracene.mol")) as f:
         mol_block = f.read()
 
-    (index, num_matches, states_searched) = at.index_search(
-            mol_block,
-            None,
-            "tree-nauty",
-            "none",  # Disable parallelism for deterministic states_searched.
-            "none",
-            "none",
-            ["int", "matchable-edges"])
-
-    assert (index, num_matches, states_searched) == (6, 466, 491)
+    assert at.index_search(
+        mol_block,
+        None,
+        "tree-nauty",
+        "none",  # Disable parallelism for deterministic states_searched.
+        "none",
+        "none",
+        ["int", "matchable-edges"],
+    ) == (6, 466, 491, True)
 
 
 def test_index_search_bad_molblock():
@@ -84,7 +83,7 @@ def test_index_search_bad_molblock():
 
 
 def test_index_search_bad_canonize():
-    with open(osp.join('data', 'checks', 'anthracene.mol')) as f:
+    with open(osp.join("data", "checks", "anthracene.mol")) as f:
         mol_block = f.read()
 
     with pytest.raises(ValueError) as e:
@@ -94,7 +93,7 @@ def test_index_search_bad_canonize():
 
 
 def test_index_search_bad_parallel():
-    with open(osp.join('data', 'checks', 'anthracene.mol')) as f:
+    with open(osp.join("data", "checks", "anthracene.mol")) as f:
         mol_block = f.read()
 
     with pytest.raises(ValueError) as e:
@@ -104,7 +103,7 @@ def test_index_search_bad_parallel():
 
 
 def test_index_search_bad_memoize():
-    with open(osp.join('data', 'checks', 'anthracene.mol')) as f:
+    with open(osp.join("data", "checks", "anthracene.mol")) as f:
         mol_block = f.read()
 
     with pytest.raises(ValueError) as e:
@@ -114,7 +113,7 @@ def test_index_search_bad_memoize():
 
 
[…]
