### 2025-12-11 b51d974
test: Add new doctest expecting a timeout in assembly::index_search
