### 2025-12-15 b956fb8
refactor: Revert Arc states_searched and API changes to index_search
