### 2025-12-15 af42b3a
feat: Clean up CLI output
