### 2025-12-15 8cf1f0a
paper: Standardize title for JOSS and add Zenodo cite for outputs
