+++ paper/paper.bib
@@ -4,6 +4,15 @@
   doi = {10.5281/zenodo.591637}
 }
 
+@software{AgentElement2025-assemblytheory,
+  title = {DaymudeLab/assembly-theory},
+  author = {AgentElement and Daymude, Josh and Mathis, Cole and Garrett-Pz and Smith, Olivia and Parkar, Devendra and Bergen, Holly},
+  year = {2025},
+  publisher = {Zenodo},
+  version = {joss-final},
+  doi = {10.5281/zenodo.16764412}
+}
+
 @article{Armstrong2020-pdbeimproved,
   title = {{{PDBe}}: Improved Findability of Macromolecular Structure Data in the {{PDB}}},
   author = {Armstrong, David R. and Berrisford, John M. and Conroy, Matthew J. and Gutmanas, Aleksandras and Anyango, Stephen and Choudhary, Preeti and Clark, Alice R. and Dana, Jose M. and Deshpande, Mandar and Dunlop, Roisin and Gane, Paul and G{\'a}borov{\'a}, Romana and Gupta, Deepti and Haslam, Pauline and Ko{\v c}a, Jaroslav and Mak, Lora and Mir, Saqib and Mukhopadhyay, Abhik and Nadzirin, Nurul and Nair, Sreenath and {Paysan-Lafosse}, Typhaine and Pravda, Lukas and Sehnal, David and Salih, Osman and Smart, Oliver and Tolchard, James and Varadi, Mihaly and {Svobodova-Va{\v r}ekov{\'a}}, Radka and Zaki, Hossam and Kleywegt, Gerard J. and Velankar, Sameer},
