### 2025-12-15 52aca0c
paper: Update assembly_go runtime on coconut_55 based on rerun
