### 2025-12-16 #9318 @jdaymude
@abhishektiwari @amcandio Thanks for your sign offs! I've updated our manuscript's title and added a citation to a [new Zenodo entry](https://zenodo.org/records/17945803) (and [GitHub release](https://github.com/DaymudeLab/assembly-theory/releases/tag/joss-final)) containing the performance evaluation outputs.

@galessiorob I believe we are ready to move forward; please let me know if anything else is needed.
