### 2025-12-23 #9318 @jdaymude
@galessiorob Thanks for taking a look! Yes, that is the "final" Zenodo entry and DOI for the `joss-paper` branch which includes the paper-specific evaluation scripts and outputs. Feel free to move forward with finalizing.
