### 2025-12-23 #9318 @galessiorob
@jdaymude What version should we use as the final? I can see the 0.6 and the JOSS final, I believe we need to use a numerical one.
