### 2025-12-23 #9318 @jdaymude
@galessiorob This is somewhat up to you and JOSS policy. The latest numerical version of our software is [v0.6.0](https://zenodo.org/records/17808464), but that doesn't include the paper-specific evaluation scripts and outputs requested by our reviewers. Those are only in [joss-final](https://zenodo.org/records/17945803), which is essentially v0.6.0 + the paper-specific stuff.
