### 2025-12-23 #9318 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
