### 2026-01-04 #9318 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] @amcandio, can you please tick the last few boxes on your checklist, provided you agree that they are satisfied?
- [ ] @jdaymude, please update the title of the Zenodo archive to match the title of the paper
- [ ] please also generate a tagged `joss-final` release on GitHub
- [ ] please fully spell out "United State of America" in affiliations in paper
- [ ] please consistently format your references in sentence case (Some Are Currently Title Case)
- [ ] please update citations to include the published version of [this paper](https://www.nature.com/articles/s44260-025-00049-9) (currently points to arXiv version)

And last but not least, some copyediting on the paper itself...
- [ ] Second sentence of summary presently reads "This is a key complexity measure of
assembly theory, a recent theoretical framework quantifying selection across diverse systems, most importantly chemistry." I think this sentence needs a bit more context and clarification – for example, what exactly is meant by the word _selection_ here? Is it in the sense of evolution, i.e. natural selection? Also, is the word _chemistry_ meant to serve as an example of _selection_, an example of a _system_, or to function in a different grammatical way?
- [ ] Bottom of first page: "Machine learning" is functioning as a [compound adjective](https://www.grammarly.com/blog/parts-of-speech/compound-adjectives/) here and hence should be hyphenated
- [ ] top of second page: naive -> naïve
- [ ] body of caption for figure 1: "anthracene" does not need to be capitalized (it can be in title of figure if your preference is for title case there)
- [ ] paragraph below figure 1: "memory-safety" does not need to be hyphenated
- [ ] same sentence: "run time" should be two words here
- [ ] Given that the paper is currently a bit long for JOSS (we typically target ≤1000 words), I suggest removing the "Functionality and Usage Examples" section as its contents are really more appropriate for the software documentation anyway (and should be there if they aren't already!)
- [ ] in Tests and Benchmarks section: in first and second bullet point, names of elements do not need to be capitalized
- [ ] second bullet point: which -> that
- [ ] paragraph after bullet list: hyphenate "ground truth" (compound adjective again)
- [ ] caption of Table 1: add a comma after `*`
- [ ] under Author contributions: hyphenate "command line"
