+++ paper/paper.md
@@ -29,11 +29,11 @@ authors:
     corresponding: true
     affiliation: "1, 3"
 affiliations:
-  - name: Biodesign Center for Biocomputing, Security and Society, Arizona State University, United States
+  - name: Biodesign Center for Biocomputing, Security and Society, Arizona State University, United States of America
     index: 1
-  - name: School of Computing and Augmented Intelligence, Arizona State University, United States
+  - name: School of Computing and Augmented Intelligence, Arizona State University, United States of America
     index: 2
-  - name: School of Complex Adaptive Systems, Arizona State University, United States
+  - name: School of Complex Adaptive Systems, Arizona State University, United States of America
     index: 3
 date: 8 December 2025
 bibliography: paper.bib
@@ -42,7 +42,7 @@ bibliography: paper.bib
 # Summary
 
 We present `assembly-theory`, an open-source, high-performance library for computing *assembly indices* of covalently bonded molecular structures.
-This is a key complexity measure of *assembly theory*, a recent theoretical framework quantifying selection across diverse systems, most importantly chemistry.
+This is a key complexity measure of *assembly theory*, a recent theoretical framework quantifying evolutionary selection across chemical, biological, and engineered systems.
 `assembly-theory` is designed for researchers and practitioners alike, providing (i) extensible, high-performance Rust implementations of assembly index calculation algorithms, (ii) comprehensive tests and benchmarks against which current and future algorithmic improvements can be evaluated, and (iii) Python bindings to support integration with existing computational pipelines.
 
 
@@ -61,115 +61,35 @@ It has also been proposed in methods to generate novel therapeutic drugs, identi
 # Statement of Need
 
 Computing MA efficiently remains a challenge.
-In general, exact MA calculation is NP-hard [@Kempes2024-assemblytheory].
+In general, exact MA calculation is NP-hard [@Kempes2025-assemblytheory].
 Previous software to compute MA have been approximate, closed-source, platform-dependent, or written in languages rarely used by the broader scientific community.
 The original software to compute a split-branch approximation of MA (an upper bound on the exact value) was written in C++ and depended on the MSVC compiler, making it difficult to deploy to non-Windows machines [@Marshall2021-identifyingmolecules].
-Machine learning methods only provide approximate MA values [@Gebhard2022-inferringmolecular].
-The more recent `assembly_go` computes MA exactly but is written in Go and implements a somewhat naive algorithm, yielding prohibitively slow performance even on mid-size molecules [@Jirasek2024-investigatingquantifying].
+Machine-learning methods only provide approximate MA values [@Gebhard2022-inferringmolecular].
+The more recent `assembly_go` computes MA exactly but is written in Go and implements a somewhat naïve algorithm, yielding prohibitively slow performance even on mid-size molecules [@Jirasek2024-investigatingquantifying].
 Finally, the latest `assemblycpp-v5` C++ implementation achieves significant performance milestones through an improved branch-and-bound approach, but lacks parallelism, has significant readability and maintenance barriers, and until recently was not publicly available for comparison or verification by the community [@Seet2025-rapidexploration].
 
[…]
