+++ paper/paper.bib
@@ -5,7 +5,7 @@
 }
 
 @software{AgentElement2025-assemblytheory,
-  title = {DaymudeLab/assembly-theory},
+  title = {{{assembly-theory}}: Open, Reproducible Calculation of Assembly Indices},
   author = {AgentElement and Daymude, Josh and Mathis, Cole and Garrett-Pz and Smith, Olivia and Parkar, Devendra and Bergen, Holly},
   year = {2025},
   publisher = {Zenodo},
@@ -13,122 +13,26 @@
   doi = {10.5281/zenodo.16764412}
 }
 
-@article{Armstrong2020-pdbeimproved,
-  title = {{{PDBe}}: Improved Findability of Macromolecular Structure Data in the {{PDB}}},
-  author = {Armstrong, David R. and Berrisford, John M. and Conroy, Matthew J. and Gutmanas, Aleksandras and Anyango, Stephen and Choudhary, Preeti and Clark, Alice R. and Dana, Jose M. and Deshpande, Mandar and Dunlop, Roisin and Gane, Paul and G{\'a}borov{\'a}, Romana and Gupta, Deepti and Haslam, Pauline and Ko{\v c}a, Jaroslav and Mak, Lora and Mir, Saqib and Mukhopadhyay, Abhik and Nadzirin, Nurul and Nair, Sreenath and {Paysan-Lafosse}, Typhaine and Pravda, Lukas and Sehnal, David and Salih, Osman and Smart, Oliver and Tolchard, James and Varadi, Mihaly and {Svobodova-Va{\v r}ekov{\'a}}, Radka and Zaki, Hossam and Kleywegt, Gerard J. and Velankar, Sameer},
-  journal = {Nucleic Acids Research},
-  volume = {48},
-  number = {D1},
-  pages = {D335-D343},
-  year = {2020},
-  doi = {10.1093/nar/gkz990}
-}
-
-@article{Barnes2021-sciencegoals,
-  title = {Science {{Goals}} and {{Objectives}} for the {{Dragonfly Titan Rotorcraft Relocatable Lander}}},
-  author = {Barnes, Jason W. and Turtle, Elizabeth P. and Trainer, Melissa G. and Lorenz, Ralph D. and MacKenzie, Shannon M. and Brinckerhoff, William B. and Cable, Morgan L. and Ernst, Carolyn M. and Freissinet, Caroline and Hand, Kevin P. and Hayes, Alexander G. and H{\"o}rst, Sarah M. and Johnson, Jeffrey R. and Karkoschka, Erich and Lawrence, David J. and Le Gall, Alice and Lora, Juan M. and McKay, Christopher P. and Miller, Richard S. and Murchie, Scott L. and Neish, Catherine D. and Newman, Claire E. and N{\'u}{\~n}ez, Jorge and Panning, Mark P. and Parsons, Ann M. and Peplowski, Patrick N. and Quick, Lynnae C. and Radebaugh, Jani and Rafkin, Scot C. R. and Shiraishi, Hiroaki and Soderblom, Jason M. and Sotzen, Kristin S. and Stickle, Angela M. and Stofan, Ellen R. and Szopa, Cyril and Tokano, Tetsuya and Wagner, Thomas and Wilson, Colin and Yingst, R. Aileen and Zacny, Kris and St{\"a}hler, Simon C.},
-  journal = {The Planetary Science Journal},
-  volume = {2},
-  number = {4},
-  pages = {130},
-  year = {2021},
-  doi = {10.3847/PSJ/abfdcf}
-}
-
 @article{Blum2009-970million,
-  title = {970 {{Million Druglike Small Molecules}} for {{Virtual Screening}} in the {{Chemical Universe Database GDB-13}}},
[…]
