### 2026-01-06 51cd1be
docs: Add CITATION.cff reflecting the JOSS publication
