+++ CITATION.cff
@@ -1,0 +1,68 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Vimal
+  given-names: Devansh
+  orcid: "https://orcid.org/0009-0006-2794-8995"
+- family-names: Parzych
+  given-names: Garrett
+  orcid: "https://orcid.org/0009-0008-4789-9603"
+- family-names: Smith
+  given-names: Olivia M.
+  orcid: "https://orcid.org/0009-0004-2299-3522"
+- family-names: Parkar
+  given-names: Devendra
+  orcid: "https://orcid.org/0009-0009-0133-8875"
+- family-names: Bergen
+  given-names: Holly
+  orcid: "https://orcid.org/0009-0004-3570-5120"
+- family-names: Daymude
+  given-names: Joshua J.
+  orcid: "https://orcid.org/0000-0001-7294-5626"
+- family-names: Mathis
+  given-names: Cole
+  orcid: "https://orcid.org/0000-0001-8424-9169"
+contact:
+- family-names: Mathis
+  given-names: Cole
+  orcid: "https://orcid.org/0000-0001-8424-9169"
+doi: 10.5281/zenodo.16764412
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Vimal
+    given-names: Devansh
+    orcid: "https://orcid.org/0009-0006-2794-8995"
+  - family-names: Parzych
+    given-names: Garrett
+    orcid: "https://orcid.org/0009-0008-4789-9603"
+  - family-names: Smith
+    given-names: Olivia M.
+    orcid: "https://orcid.org/0009-0004-2299-3522"
+  - family-names: Parkar
+    given-names: Devendra
+    orcid: "https://orcid.org/0009-0009-0133-8875"
+  - family-names: Bergen
+    given-names: Holly
+    orcid: "https://orcid.org/0009-0004-3570-5120"
+  - family-names: Daymude
+    given-names: Joshua J.
+    orcid: "https://orcid.org/0000-0001-7294-5626"
+  - family-names: Mathis
+    given-names: Cole
+    orcid: "https://orcid.org/0000-0001-8424-9169"
+  date-published: 2026-01-06
+  doi: 10.21105/joss.09318
+  issn: 2475-9066
+  issue: 117
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 9318
+  title: "assembly-theory: Open, Reproducible Calculation of Assembly
+    Indices"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.09318"
+  volume: 11
+title: "assembly-theory: Open, Reproducible Calculation of Assembly
+  Indices"
