### 2026-01-06 c8eb357
docs: Update the three READMEs with the JOSS citation
